Word vectors data type: wordvec and embed.

PsychWordVec uses two types of word vectors data: wordvec (data.table, with two variables word and vec) and embed (matrix, with dimensions as columns and words as row names). Note that matrix manipulation makes embed much faster than wordvec. Users are suggested to reshape data to embed before using the other functions.

Usage

as_embed(x, normalize = FALSE)

as_wordvec(x, normalize = FALSE)

Arguments

x

Object to be reshaped. See examples.

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See normalize.

Value

A wordvec (data.table) or embed (matrix).

Functions

as_embed(): From wordvec (data.table) to embed (matrix).

as_wordvec(): From embed (matrix) to wordvec (data.table).

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

See also

Examples

d = head(demodata, 10)

embed = as_embed(d)
embed
#> # embed (matrix): 10 × 300 (NOT normalized)
#>             dim1 ...     dim300
#>  1: in    0.0703 ... <300 dims>
#>  2: for  -0.0118 ... <300 dims>
#>  3: that -0.0157 ... <300 dims>
#>  4: is    0.0070 ... <300 dims>
#>  5: on    0.0267 ... <300 dims>
#>  6: with -0.0249 ... <300 dims>
#>  7: said -0.0091 ... <300 dims>
#>  8: was   0.0260 ... <300 dims>
#>  9: the   0.0801 ... <300 dims>
#> 10: at   -0.0586 ... <300 dims>

wordvec = as_wordvec(embed)
wordvec
#> # wordvec (data.table): 10 × 2 (NOT normalized)
#>     word                      vec
#>  1:   in [ 0.0703, ...<300 dims>]
#>  2:  for [-0.0118, ...<300 dims>]
#>  3: that [-0.0157, ...<300 dims>]
#>  4:   is [ 0.0070, ...<300 dims>]
#>  5:   on [ 0.0267, ...<300 dims>]
#>  6: with [-0.0249, ...<300 dims>]
#>  7: said [-0.0091, ...<300 dims>]
#>  8:  was [ 0.0260, ...<300 dims>]
#>  9:  the [ 0.0801, ...<300 dims>]
#> 10:   at [-0.0586, ...<300 dims>]