Reshape word vectors data.

Reshape word vectors data from dense (data.table) to plain (matrix) or vice versa.

Usage

data_wordvec_reshape(
  data,
  to = c("plain", "dense"),
  normalize = FALSE,
  verbose = TRUE
)

Arguments

data

Data to be reshaped. See examples.

to

Options include:

"plain" (default) reshapes the data from data.table (with two variables word and vec, loaded by data_wordvec_load) to matrix (with dimensions as columns and words as row names).

"dense" just does the reverse.

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See data_wordvec_normalize.

verbose

Print information to the console? Defaults to TRUE.

Value

A data.table (dense) or matrix (plain) of word vectors.

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

See also

Examples

d = head(demodata, 10)
d
#>     word
#>  1:   in
#>  2:  for
#>  3: that
#>  4:   is
#>  5:   on
#>  6: with
#>  7: said
#>  8:  was
#>  9:  the
#> 10:   at
#>                                                                                         vec
#>  1:              0.05295587, 0.06545975, 0.06619558, 0.04707222, 0.05222079,-0.08200884,...
#>  2:       -0.008512173,-0.034224283, 0.032284114, 0.045868184,-0.013142564,-0.046220811,...
#>  3:             -0.01236052,-0.02222963, 0.06553974, 0.03947722,-0.08662021, 0.02491257,...
#>  4:        0.003746262,-0.038919677, 0.091331742, 0.012000260,-0.070574262, 0.105342762,...
#>  5:              0.01668099,-0.05667031, 0.01736675, 0.12735656, 0.00388493,-0.05636580,...
#>  6:             -0.01596967, 0.01409130,-0.02270204, 0.08767797, 0.01048975,-0.01792692,...
#>  7:       -0.004531971,-0.022021472, 0.049639883,-0.037960116,-0.028226893, 0.030659825,...
#>  8:  0.0126005706,-0.0009168986, 0.0899195443,-0.0250828942, 0.0024846400,-0.0532416633,...
#>  9:              0.07468524, 0.09791024, 0.04645094, 0.04986632,-0.06284517,-0.11248299,...
#> 10:             -0.03536007,-0.02268949, 0.04390528, 0.06571068, 0.04007442, 0.02931927,...

d.plain = data_wordvec_reshape(d, to="plain")
d.plain
#>              [,1]          [,2]        [,3]        [,4]        [,5]        [,6]
#> in    0.052955865  0.0654597520  0.06619558  0.04707222  0.05222079 -0.08200884
#> for  -0.008512173 -0.0342242830  0.03228411  0.04586818 -0.01314256 -0.04622081
#> that -0.012360524 -0.0222296329  0.06553974  0.03947722 -0.08662021  0.02491257
#> is    0.003746262 -0.0389196767  0.09133174  0.01200026 -0.07057426  0.10534276
#> on    0.016680988 -0.0566703074  0.01736675  0.12735656  0.00388493 -0.05636580
#> with -0.015969667  0.0140912978 -0.02270204  0.08767797  0.01048975 -0.01792692
#> said -0.004531971 -0.0220214718  0.04963988 -0.03796012 -0.02822689  0.03065983
#> was   0.012600571 -0.0009168986  0.08991954 -0.02508289  0.00248464 -0.05324166
#> the   0.074685235  0.0979102375  0.04645094  0.04986632 -0.06284517 -0.11248299
#> at   -0.035360065 -0.0226894859  0.04390528  0.06571068  0.04007442  0.02931927
#>               [,7]        [,8]        [,9]       [,10]        [,11]       [,12]
#> in   -0.0614145542 -0.11620963 0.015629482  0.09929300 -0.085686495 -0.02813337
#> for  -0.0009480451 -0.05221907 0.046573437  0.06245102 -0.122784845 -0.02875568
#> that -0.0111626981 -0.07052257 0.092369148  0.09275220 -0.056340973 -0.06055691
#> is    0.0599364722 -0.05734172 0.038141198  0.01109212 -0.065385291 -0.03139527
#> on    0.0140914507 -0.10054456 0.082872681  0.03808492 -0.009825890  0.05514716
#> with -0.0155784738 -0.08955634 0.036793338  0.03992481 -0.075465688 -0.06231390
#> said  0.1275066591 -0.07884005 0.008273563 -0.04818022 -0.031633596 -0.07543335
#> was  -0.0039636963 -0.04283012 0.046852935  0.02342647  0.005353583 -0.17605276
#> the   0.0327884579 -0.11066151 0.040985572  0.02812051 -0.053054142 -0.07104228
#> at    0.0766130724 -0.03506497 0.041548107  0.05451318 -0.120223860 -0.02460431
#>             [,13]        [,14]       [,15]       [,16]        [,17]       [,18]
#> in    0.052220786  0.058840269 -0.07759535 -0.07355015  0.033281186  0.07722781
#> for   0.051513099 -0.018700045  0.01314256  0.09879251  0.104437427 -0.02434496
#> that -0.054041870  0.060556911 -0.10884985  0.00510214  0.008623402 -0.01188171
#> is    0.053709175  0.005708134  0.07628293  0.13803567 -0.019330180  0.09859683
#> on    0.008607248  0.028944787 -0.03488575 -0.04174147  0.007655047  0.08531059
#> with -0.112102549 -0.013230031 -0.07640519  0.06356636  0.038672349  0.09018289
#> said  0.037230037  0.069106826 -0.11485362  0.03102511  0.088086884  0.10804021
#> was   0.039753756 -0.010293301  0.07666813  0.02141482 -0.005708808  0.10317096
#> the   0.012010745  0.046450937 -0.07923940 -0.05920221  0.005863609 -0.04030287
#> at    0.026962095 -0.052450502 -0.03285504  0.02165815 -0.033444633  0.06836235
#>              [,19]       [,20]         [,21]       [,22]        [,23]
#> in   -0.0457850724 -0.02721377 -0.0342007886  0.03567170 -0.090834312
#> for  -0.0705657703 -0.08679598 -0.0575113629  0.04516221 -0.048338015
#> that -0.0007543318 -0.02376263 -0.0001240213  0.03008693 -0.018972113
#> is   -0.0417737438 -0.01206509 -0.0640876493  0.09444512  0.026335956
#> on    0.0962789960 -0.02879253 -0.0245269739 -0.09627900 -0.102981842
#> with -0.0403942412 -0.03131342 -0.0291214596  0.02191965  0.007241566
#> said -0.1245868418 -0.02092660 -0.1002535366  0.02409011  0.059616797
#> was  -0.0278040282  0.02744929 -0.0520587091  0.08991954  0.037387363
#> the   0.0188993432  0.01240992 -0.0182157063  0.08652530 -0.160300267
#> at    0.0444942761 -0.02872968 -0.0842741986 -0.06335290  0.020037209
#>              [,24]        [,25]       [,26]        [,27]        [,28]
#> in   -4.817559e-02  0.001700625  0.02794885 -0.002194695  0.088628320
#> for   5.363030e-02  0.016406527  0.02416865 -0.130546964  0.037576403
#> that -3.602739e-02  0.074355453 -0.04331011 -0.050975092  0.004791302
#> is    9.146193e-03  0.042292907 -0.02425984 -0.100153790  0.100672422
#> on    6.733421e-02  0.020718170 -0.03183883  0.023155455  0.063373153
#> with -2.896498e-02  0.125253700 -0.02458105 -0.084546504  0.040707837
#> said  1.387000e-02  0.055966901 -0.06716028 -0.054506743 -0.029443359
#> was   7.026971e-05  0.073828750 -0.03170860 -0.073828750  0.109323192
#> the  -1.231106e-03  0.061023689  0.05442048 -0.076962781  0.080150600
#> at    5.561634e-03 -0.028729676 -0.01804815 -0.053629090  0.044789376
#>              [,29]       [,30]        [,31]         [,32]        [,33]
#> in    0.0465209053 0.048175588  0.061047014 -0.0518532458 -0.016088907
#> for   0.0100115584 0.067037337  0.002536310 -0.0065712820 -0.070213144
#> that  0.0006703427 0.048675989 -0.042734742  0.0110669348  0.017439116
#> is   -0.0123244049 0.033470858  0.051892902 -0.0101191585 -0.042033591
#> on    0.0688579802 0.012796682  0.004227498  0.0007381741 -0.007807299
#> with -0.0120558097 0.085172413  0.036636861  0.0142477751 -0.030060965
#> said -0.1031733538 0.041123625 -0.061320149  0.0231168393  0.028104798
#> was   0.0392807680 0.043066609  0.035494442 -0.0501657885 -0.032418567
#> the  -0.0029742403 0.054648052 -0.032560890 -0.0114987171 -0.044857028
#> at    0.0324132926 0.006740222  0.049209233 -0.0556917709 -0.134957120
#>             [,34]        [,35]        [,36]       [,37]        [,38]
#> in    0.041556105 -0.064356379  0.051853246 -0.09635118 -0.025007020
#> for   0.049043268 -0.006350890  0.031930765 -0.09667531 -0.071976999
#> that -0.035261284  0.087386320 -0.030278456  0.04024411  0.019738219
#> is    0.080953268  0.092369536  0.053968491 -0.08666140  0.060974266
#> on   -0.077998111  0.009254944 -0.016757114 -0.01340569  0.009407196
#> with  0.058242919 -0.112728458  0.015813190 -0.03115695  0.003542544
#> said  0.016425031  0.036986843 -0.015816549  0.02141349  0.025671866
#> was   0.085660231  0.103170958  0.068149503 -0.01656427  0.008636874
#> the  -0.002817554  0.052598073  0.013946939 -0.06739933 -0.048727551
#> at   -0.100186650 -0.025341159 -0.006188042 -0.04478938 -0.077202668
#>             [,39]         [,40]       [,41]        [,42]       [,43]
#> in    0.074285981  0.1323904167  0.08347975 -0.026110393 -0.03548793
#> for   0.023992333  0.0201994301  0.11219955 -0.012789937  0.01032011
#> that  0.013318941  0.0494420946  0.10808296  0.106549959  0.05135893
#> is    0.053449859 -0.0492981509  0.05812020  0.031265612 -0.01148110
#> on    0.086529232  0.0303156817 -0.04783468 -0.072818412  0.06672520
#> with  0.035227924  0.0077109982  0.07734406  0.019101138 -0.10208223
#> said -0.077379891 -0.0008058277 -0.01429609 -0.046963259  0.03966347
#> was   0.028868735  0.0001386009  0.02851399  0.004495808  0.08092744
#> the   0.090169192  0.0400753000 -0.03301603 -0.068309598  0.03051184
#> at    0.002725901  0.0539235866 -0.01804815 -0.104312012 -0.04302119
#>             [,44]        [,45]        [,46]        [,47]        [,48]
#> in   -0.006389771  0.027029996  0.077595346  0.020318628 -0.021605770
#> for  -0.079386492 -0.061745049 -0.052924328 -0.017817756  0.124902049
#> that -0.050592039 -0.018876349 -0.010492355 -0.029128512  0.003377617
#> is    0.033730705  0.022313905 -0.004702754  0.017124934  0.056563242
#> on    0.026050121  0.007997614 -0.005902897 -0.018052507 -0.024069593
#> with  0.060748481 -0.034914328  0.006849732  0.072647172 -0.025989989
#> said -0.098793379 -0.103173354 -0.038690195 -0.057670252 -0.047450145
#> was  -0.001308955  0.056081529  0.050638776  0.026029355  0.008992100
#> the  -0.057608302  0.007343734  0.003315592 -0.008197114  0.036431413
#> at   -0.045967964 -0.006482538  0.052450502 -0.033297385  0.136135708
#>             [,49]       [,50]        [,51]       [,52]        [,53]       [,54]
#> in   -0.003861428  0.08016963  0.045049240  0.07097586  0.025374560 -0.02041051
#> for   0.044633269  0.06456751 -0.017553286  0.10232094 -0.023816020  0.01984680
#> that -0.012360524  0.01494770  0.085086432  0.03506976 -0.035261284 -0.07473851
#> is    0.034249337 -0.06304986  0.016216798  0.03554698  0.064866128 -0.04410918
#> on    0.151730661  0.00594096  0.013786946  0.13893398  0.005712582 -0.02833515
#> with  0.154688755  0.08204158 -0.009629128  0.02880850 -0.082041584 -0.05855652
#> said -0.007056599  0.07543335  0.086626726 -0.01155817  0.075920231 -0.05304659
#> was   0.051585237  0.02449117 -0.007986031 -0.06199727  0.105063878  0.02496465
#> the   0.035748708  0.04144164  0.065121780  0.01178318 -0.004155916 -0.03096698
#> at    0.027698336  0.10254383 -0.010166155  0.08898916  0.007035322 -0.01009253
#>            [,55]        [,56]        [,57]        [,58]        [,59]
#> in   -0.07097586 0.0007757501 -0.036407534  0.025926623  0.061047014
#> for  -0.11219955 0.0056897157 -0.051160472  0.031578139  0.004344243
#> that  0.06822268 0.0643897959  0.005365882 -0.103483965  0.002143684
#> is    0.09133174 0.0417737438  0.050595792 -0.004135236  0.012324405
#> on   -0.02208906 0.0877478746 -0.115168890  0.048444315  0.025897869
#> with  0.08955634 0.0989507564 -0.052293574  0.010567992  0.024424574
#> said  0.02737522 0.0715402563 -0.054506743  0.006935002  0.054993629
#> was   0.02318997 0.0738287499  0.083293838  0.038097329 -0.028513995
#> the  -0.03984773 0.0910794642 -0.020150968 -0.035293572  0.011100473
#> at   -0.06718376 0.0386013342 -0.044789376  0.067773358 -0.004751769
#>            [,60]        [,61]       [,62]       [,63]        [,64]       [,65]
#> in   -0.08531820 -0.066930665 0.027029996 -0.10959014 -0.183876123 -0.04633638
#> for  -0.04039886 -0.106554631 0.020552057 -0.09596933 -0.127724507 -0.06597946
#> that -0.05940775  0.017630642 0.134912364 -0.03813575  0.030086929 -0.06937262
#> is    0.01245459 -0.008951706 0.082509693 -0.05837952 -0.093926493 -0.06175275
#> on   -0.05301438 -0.061850006 0.036409518 -0.06032686 -0.126137297 -0.00875950
#> with -0.09080945  0.003718261 0.043838667 -0.10897108 -0.066697826 -0.06200094
#> said -0.04939669 -0.091979974 0.022508357 -0.04720695  0.007756278 -0.01423479
#> was  -0.02082358 -0.035257948 0.088026624 -0.03904379 -0.074775210 -0.05655452
#> the  -0.01297884 -0.105652216 0.086981373 -0.03665991 -0.108384898  0.02174488
#> at   -0.06070122 -0.169727589 0.009871055 -0.05952264 -0.168549001 -0.05097742
#>             [,66]        [,67]         [,68]        [,69]      [,70]
#> in    0.039901422  0.042843247  0.1353329950  0.045968842 0.06545975
#> for  -0.036694114 -0.018788201 -0.1072606067 -0.058216616 0.10867184
#> that -0.013510468  0.017151826  0.1050169619  0.008383994 0.03909417
#> is    0.012454594 -0.005643305  0.0280220397 -0.071093425 0.04229291
#> on   -0.001475724  0.092013435  0.1249186543  0.037170780 0.09627900
#> with  0.067950286 -0.016361502  0.0389853039 -0.090809445 0.11460747
#> said  0.066673396  0.021048199  0.0001210984 -0.067160281 0.09003343
#> was   0.040937195 -0.029933927 -0.0218882878  0.004466246 0.06483665
#> the  -0.014914104  0.024591343  0.1001877837 -0.004354571 0.08971312
#> at   -0.033002889  0.061585316  0.1160984988  0.047146553 0.12729599
#>             [,71]         [,72]       [,73]       [,74]       [,75]
#> in    0.093408597 -3.033936e-02  0.01700851  0.13386133 -0.02234085
#> for  -0.031401826  5.715874e-02  0.02399233  0.06527348  0.01640653
#> that  0.029895403 -4.120175e-03  0.04810062 -0.03928570 -0.08317038
#> is    0.039178992  2.335170e-02  0.06123358 -0.01096247  0.03969816
#> on    0.084091947  3.294640e-03  0.07799811  0.05331888 -0.01729062
#> with  0.060121930  8.705142e-02 -0.01471721  0.03788932  0.03272236
#> said -0.042826976 -2.993024e-02  0.06521324  0.05523682 -0.07932693
#> was   0.009346841  3.123513e-02  0.05276868  0.07666813 -0.01141762
#> the   0.026071469 -5.032146e-02  0.07969453 -0.03438237 -0.01889934
#> at    0.068067854  8.810748e-05  0.07955985  0.01023978  0.06423699
#>             [,76]       [,77]       [,78]        [,79]      [,80]        [,81]
#> in   -0.022340850  0.08826078 0.023444223 -0.072446775 0.05001404  0.003539831
#> for  -0.045162209  0.11855044 0.062098398 -0.008952956 0.14183752 -0.044985895
#> that  0.043693160  0.12111461 0.134145473  0.037752699 0.09965108  0.064006743
#> is   -0.006130054  0.04281154 0.022184248  0.042552222 0.18681466  0.051374271
#> on   -0.036256642  0.11455988 0.004913257 -0.095669363 0.07738848 -0.049967462
#> with -0.100203216  0.09018289 0.156567766  0.034288418 0.12462779  0.027085972
#> said  0.019223251  0.02798320 0.139186925  0.071053371 0.02360323  0.083706909
#> was   0.059867370  0.08234738 0.016919008  0.062943730 0.12872684  0.093705386
#> the  -0.079694531  0.11658201 0.134797719  0.024932696 0.14026308  0.049182687
#> at   -0.013922792 -0.04419978 0.076613072 -0.057754451 0.03005612  0.038306838
#>             [,82]        [,83]        [,84]        [,85]        [,86]
#> in   -0.060311181  0.047439755 -0.015537597 -0.041188565 -0.102234826
#> for   0.016230214 -0.021786972  0.015347925  0.002403352 -0.040046234
#> that -0.079720550  0.034494393 -0.010636000 -0.105016962 -0.123414495
#> is   -0.113127016  0.087699196 -0.041514428  0.037103404 -0.000741813
#> on   -0.089576150  0.093232078  0.009102068  0.067334209 -0.125528288
#> with -0.079849618 -0.004951482 -0.024737529  0.078283562 -0.115234020
#> said  0.041123625  0.069593213  0.065213238 -0.027740008 -0.072999916
#> was  -0.042830115  0.041173689 -0.014375244  0.027922275  0.012186222
#> the  -0.173961814  0.076051576 -0.009904808 -0.034838436 -0.068309598
#> at   -0.004622625  0.082506618  0.055102778  0.002596757 -0.027551087
#>            [,87]       [,88]       [,89]        [,90]        [,91]        [,92]
#> in   -0.04780805  0.06288547 -0.04817559  0.016180792  0.058104436 -0.027948845
#> for  -0.05292433  0.02152250  0.03598886  0.012613624  0.075505432  0.028049706
#> that  0.01906788  0.16404088 -0.08010439 -0.073588562  0.038135752  0.059023914
#> is   -0.04852020  0.06901784  0.13388396 -0.008562201  0.049557998 -0.077839353
#> on   -0.09384171  0.03275297  0.04813919  0.057279941 -0.023765088  0.092623069
#> with -0.16408253 -0.01268172 -0.04728309  0.096445836 -0.008063072  0.044465217
#> said -0.07932693  0.15476028 -0.01143657 -0.022265164  0.167413818 -0.014113197
#> was  -0.04921884  0.09654525  0.05063878 -0.011772357  0.009761190 -0.017037255
#> the  -0.07013108  0.03369966 -0.12295765  0.005749825  0.047361209  0.042124345
#> at    0.08073843 -0.01038703 -0.07366630  0.024899414 -0.059522636 -0.009723807
#>              [,93]        [,94]        [,95]        [,96]        [,97]
#> in   -0.0253745600 -0.138274820 -0.054795071  0.011951824  0.070240783
#> for   0.0613924225 -0.179237606  0.050101870 -0.107965860  0.042163438
#> that  0.0027669299 -0.096968139 -0.018972113 -0.001036127  0.030086929
#> is   -0.0008029223 -0.080434105 -0.014270336 -0.083548019  0.014011020
#> on    0.0341244855 -0.094450720  0.008911752 -0.065505933  0.011882544
#> with -0.0090808163 -0.029121460 -0.051354068 -0.062626850  0.035853833
#> said  0.0462331803  0.009794021 -0.027618411 -0.072999916  0.008516757
#> was   0.0321815888 -0.030052174  0.014079627 -0.047562901 -0.088026624
#> the   0.0093927796 -0.140263084 -0.056014393  0.053736847 -0.006460509
#> at   -0.0097974308 -0.060701224 -0.106079592 -0.030350612  0.088399560
#>             [,98]       [,99]      [,100]        [,101]       [,102]
#> in   -0.046336382 -0.01071062 -0.00259688  0.0083660506 -0.119151450
#> for   0.069154542 -0.02416865  0.04551484  0.0154360816 -0.105143402
#> that  0.005964795  0.01389352  0.03430287 -0.0770383942 -0.045609210
#> is    0.045666137  0.03814120  0.04099527 -0.0207569484  0.028930176
#> on   -0.039608689  0.03336260  0.02178456  0.0871382414 -0.083482314
#> with  0.011742855 -0.06012193  0.10583961  0.0416467013 -0.133395652
#> said  0.103660239 -0.03211998 -0.07348680 -0.0564532881 -0.037230037
#> was   0.104117418 -0.02792228 -0.02188829  0.0217700409 -0.009465088
#> the   0.014800320 -0.01992340  0.09654483 -0.0002723356 -0.043718255
#> at    0.023572975 -0.07101463 -0.02872968  0.0580489476  0.002209929
#>           [,103]       [,104]       [,105]        [,106]       [,107]
#> in   -0.01287143  0.004665798 -0.006573541 -0.0606787213 -0.011492399
#> for   0.03881132 -0.065626109 -0.018347418  0.0328130543  0.003836981
#> that  0.01106693  0.032195290 -0.027787041 -0.0180136951 -0.102717074
#> is   -0.06798004  0.048520204 -0.098077670 -0.0199790012 -0.014854327
#> on    0.13527805 -0.121262727 -0.036409518 -0.0840919472 -0.165745985
#> with -0.09018289  0.015187280  0.012995315  0.0385158718  0.047909643
#> said -0.01058490  0.061076955  0.072026643  0.0345534131 -0.023846918
#> was  -0.10127755  0.057974449 -0.043776576  0.0246094218  0.004377561
#> the  -0.01525546 -0.073318894 -0.064666643  0.0152554558 -0.029373071
#> at    0.08722097 -0.072782208 -0.016427820  0.0005847681  0.043021191
#>           [,108]       [,109]       [,110]       [,111]       [,112]
#> in   -0.06619558  0.002620228 -0.012135594 -0.009285653  0.073550148
#> for  -0.08362090 -0.014113010  0.087501960  0.023287080  0.068449289
#> that -0.11344884  0.022709234 -0.096202033 -0.055957920 -0.005605290
#> is   -0.04774173 -0.062012061 -0.052412065  0.025557478 -0.086142239
#> on   -0.06489692  0.022089065 -0.134668418  0.051491233  0.028182901
#> with -0.02364155  0.016987411 -0.001594915  0.002387562  0.049162103
#> said -0.08906016 -0.072999916  0.024576499  0.017763592 -0.059616797
#> was  -0.04732592 -0.038097329 -0.066256583 -0.055135069 -0.002736157
#> the  -0.01280817 -0.034154799 -0.082882350 -0.044857028 -0.012296138
#> at   -0.05009333 -0.035212817 -0.050387821  0.083685206  0.016354196
#>           [,113]        [,114]       [,115]       [,116]       [,117]
#> in   -0.10517665 -0.0647246723 -0.020225989  0.040636502  0.100028080
#> for  -0.04657344  0.0164065271  0.087148611  0.043574667  0.087148611
#> that -0.07588845  0.0459930477  0.081637385  0.020696636  0.005940461
#> is   -0.05786088  0.0451469738 -0.024259836  0.084066650 -0.020238317
#> on   -0.04356912 -0.0269642589  0.016833864 -0.056365803  0.003427549
#> with -0.02990449 -0.0519806189 -0.022859159  0.127759261  0.098324206
#> said  0.00158923  0.0102201078 -0.068133554 -0.120206867 -0.004775663
#> was  -0.01975839 -0.0009983145 -0.027449287  0.001767404  0.040227228
#> the  -0.06694326  0.0054933444 -0.043035550  0.037115050  0.093812147
#> at   -0.07130912 -0.0515664098 -0.009650183  0.005267137  0.091346332
#>            [,118]      [,119]     [,120]       [,121]      [,122]      [,123]
#> in    0.084950662  0.09120261 0.06435638 -0.005355688  0.03364948 -0.10959014
#> for   0.035282885  0.06739069 0.04833802  0.021169876 -0.02469831 -0.08044509
#> that  0.028362406  0.03161993 0.04139327 -0.160208776 -0.02625404 -0.02280500
#> is   -0.043590017  0.10793751 0.04592545  0.036844088  0.01712493 -0.08510444
#> on    0.031077567 -0.02224194 0.03732303 -0.093232078 -0.01378695 -0.02071817
#> with  0.047596688  0.02552056 0.04070784 -0.068263241 -0.02505048 -0.07139471
#> said  0.015816549  0.02920017 0.04234009 -0.107066942  0.11582689 -0.02956496
#> was  -0.034311488  0.12778038 0.06034084 -0.007808661  0.04425005 -0.11594890
#> the  -0.045994868  0.07058714 0.03620384 -0.155746108 -0.08971312 -0.09426728
#> at   -0.003775345  0.15322614 0.03418148 -0.001169536  0.01569098  0.04449428
#>            [,124]      [,125]       [,126]       [,127]      [,128]
#> in   -0.002413110 -0.08862832 -0.049278961  0.053691698 -0.07097586
#> for   0.038635005 -0.01852373  0.012878094  0.044985895 -0.01870004
#> that  0.024912574 -0.09696814 -0.052891926  0.012456287 -0.06783963
#> is    0.050336477 -0.01310288  0.028800518  0.014854327  0.02387086
#> on    0.109684687 -0.04143634 -0.011425787  0.007046037 -0.02635463
#> with -0.013855940 -0.03554088 -0.057929964 -0.002524800 -0.07296077
#> said  0.063753579 -0.05742656 -0.100253537  0.043800248 -0.02749681
#> was  -0.005087527 -0.03289156  0.068149503  0.113582506 -0.03099863
#> the   0.027096458 -0.05396441 -0.017988138 -0.040075300 -0.07832819
#> at   -0.035212817 -0.02843518 -0.001187641 -0.045378368 -0.03359188
#>            [,129]      [,130]       [,131]       [,132]       [,133]
#> in   -0.022800275  0.09046677  0.060311181 -0.071344156 -0.122093275
#> for   0.105143402  0.04516221  0.077975263 -0.117845183 -0.070565770
#> that  0.009821228 -0.04944209 -0.094668251  0.018397533 -0.103483965
#> is    0.090293948  0.03866036 -0.019330180 -0.005448819 -0.009081364
#> on   -0.048139186  0.01089228 -0.065505933 -0.065201429 -0.029553796
#> with  0.079536022  0.07859716  0.035071446 -0.046657183 -0.114607469
#> said  0.015025671  0.01320072  0.006722208 -0.066186511 -0.016060241
#> was   0.094651846  0.02437293 -0.060577336  0.009997684 -0.057737956
#> the  -0.018557058  0.04804485  0.007912654 -0.033699663 -0.139351879
#> at    0.017090437  0.06040673 -0.010018906 -0.152047556 -0.015322856
#>           [,134]        [,135]      [,136]       [,137]      [,138]
#> in   -0.05847273  1.599702e-02 -0.06141455  0.002965173 -0.11841637
#> for  -0.07656403 -6.174505e-02 -0.06421488  0.073035601 -0.05751136
#> that -0.09275220 -9.533938e-03 -0.08623638  0.074738506 -0.03296140
#> is   -0.05708241 -3.730320e-04 -0.03917899  0.134921759  0.03009816
#> on   -0.01843314 -3.838942e-02 -0.03168658 -0.015995853 -0.05941272
#> with  0.04571768 -4.289980e-02  0.01964945 -0.004970721  0.02505048
#> said -0.03990666 -7.361089e-03 -0.06107695  0.026888332  0.07543335
#> was  -0.05679150 -4.361568e-05  0.01786547  0.070988884  0.01200885
#> the  -0.01730543 -3.392723e-02 -0.07149742 -0.036887481 -0.05737980
#> at   -0.08545339 -1.675852e-03 -0.04685206  0.050387821 -0.02976102
#>            [,139]       [,140]      [,141]       [,142]       [,143]
#> in   -0.073918441  0.029971821  0.02960428 -0.006849196  0.077595346
#> for   0.006086421  0.017376973  0.09455810  0.037047463  0.058922592
#> that  0.014373122  0.040627166  0.01173806 -0.124947492 -0.017056063
#> is    0.018681360 -0.004573097  0.09859683  0.011416268  0.140111261
#> on   -0.050577095 -0.063373153  0.12613730  0.074037055 -0.001761509
#> with  0.027712522 -0.033348913 -0.01589143  0.130890732 -0.043525712
#> said  0.045016714  0.033580141 -0.01727671 -0.014478486  0.002661673
#> was  -0.072408817  0.001471302 -0.01502609  0.051822215  0.123994054
#> the  -0.018671775  0.038708959  0.03438237 -0.067399326  0.005521324
#> at   -0.085453390  0.120223860  0.03830684  0.014880508  0.001390408
#>            [,144]       [,145]       [,146]       [,147]      [,148]
#> in    0.051117413 -0.032178566  0.047808048 -0.036958844  0.01572137
#> for   0.067743313 -0.042339751 -0.069860517 -0.020463900 -0.10514340
#> that -0.004024411  0.028170879 -0.002383092 -0.061323802 -0.04024411
#> is   -0.012649081 -0.052671381 -0.021924932 -0.036844088 -0.06045564
#> on   -0.037628161  0.014015325  0.081654038  0.050271966 -0.09627900
#> with  0.023328591  0.042273252  0.019884166  0.005440794 -0.12525370
#> said -0.036013571  0.004379975  0.063753579 -0.011436574 -0.11971998
#> was  -0.029105714 -0.100331092  0.076668131 -0.104590406 -0.08944607
#> the  -0.058746143  0.006887665 -0.014800320  0.015027888 -0.01354776
#> at   -0.052156006 -0.027846187  0.008177098  0.042726695 -0.09841847
#>            [,149]       [,150]       [,151]      [,152]        [,153]
#> in   -0.125770933  0.070240783  0.070608323 0.005171165  0.0404527319
#> for  -0.106554631  0.105143402 -0.012701781 0.023816020 -0.0617450490
#> that -0.005821150  0.068605734 -0.018780586 0.034685920 -0.0893031546
#> is    0.026595272 -0.031265612  0.024519683 0.046444084  0.0560446108
#> on   -0.051491233  0.099934923  0.036104390 0.060935868 -0.0131011867
#> with -0.001614795  0.118364849  0.092061905 0.051667023 -0.0078674756
#> said  0.014539284  0.026279850  0.033580141 0.133346792 -0.0596167972
#> was  -0.083767310  0.003875011 -0.014730470 0.047562901  0.0022626847
#> the   0.007201038  0.094267283 -0.005208884 0.013320194 -0.0710422805
#> at   -0.076024080  0.034475973 -0.020774053 0.128474582 -0.0009667683
#>            [,154]       [,155]       [,156]      [,157]        [,158]
#> in    0.039533129 -0.018387537 -0.024455710 -0.04633638 -0.0041830253
#> for  -0.007938432 -0.026814790 -0.009879323  0.02593322 -0.0059541856
#> that  0.016863751 -0.003006338 -0.034111340 -0.08163739 -0.1456441279
#> is    0.056563242  0.014854327  0.050336477  0.06175275 -0.0918509047
#> on   -0.028182901 -0.045701901  0.002723070 -0.05667031  0.0119586703
#> with  0.045404723 -0.019962405 -0.013621224 -0.08141503 -0.1183648490
#> said  0.114853619  0.032363176  0.011801862 -0.02056181  0.0006309078
#> was   0.124940514  0.051822215  0.061523797  0.03076214 -0.0634167173
#> the   0.052598073 -0.012067637  0.028576583 -0.02322500 -0.0919906692
#> at    0.044494276 -0.016943189 -0.053334594 -0.07602408  0.0086924665
#>            [,159]      [,160]       [,161]       [,162]        [,163]
#> in    0.072446775  0.02850091  0.009193768 -0.033097416 -0.0056313432
#> for   0.036341487 -0.06844929  0.034577632  0.014995298  0.0221403211
#> that -0.035261284  0.03526128 -0.034878231  0.014947701 -0.0164806982
#> is   -0.018551702 -0.11053226  0.031654585  0.055266132 -0.0009527727
#> on   -0.010359397 -0.03138207  0.009178194 -0.002589537  0.0216323080
#> with  0.015813190  0.05573800 -0.003092351  0.070142251 -0.0358538331
#> said -0.016789821  0.02177828  0.020683409  0.041123625  0.0367431514
#> was   0.031708601 -0.04803589  0.028632242 -0.017983715  0.0489823496
#> the   0.030284276 -0.02618525 -0.075596440  0.019240695  0.0150278878
#> at   -0.009429311  0.04567347  0.041548107 -0.046852056 -0.0166486924
#>           [,164]      [,165]       [,166]       [,167]       [,168]
#> in    0.07943455  0.01535383  0.109590142  0.061782094  0.004344201
#> for   0.09385285  0.03810534  0.013584070 -0.012701781  0.025227248
#> that  0.01058812  0.01197747 -0.023859178  0.036602755  0.080487441
#> is    0.03113595 -0.01582729 -0.019979001  0.025816793 -0.033990021
#> on    0.03579988  0.04996746  0.003884930  0.039913193  0.015310093
#> with  0.12588025  0.03663686  0.067950286  0.050414563 -0.060434885
#> said  0.02336003  0.02907857 -0.005018856  0.001452184 -0.025184981
#> was   0.07430174 -0.07430174 -0.036677881  0.028395748 -0.024491175
#> the  -0.03916410 -0.03256089 -0.035066004  0.051232665  0.012808166
#> at    0.06246941  0.02489941  0.070720131 -0.007661126  0.025046663
#>            [,169]       [,170]      [,171]       [,172]       [,173]
#> in    0.003447946 -0.069872490 -0.10444157 -0.043210787 -0.038798049
#> for   0.013231443 -0.007145023 -0.13336942 -0.064920856 -0.020993562
#> that -0.010875408  0.006467944 -0.04139327  0.015426518 -0.059790805
#> is    0.042292907  0.036844088 -0.05578476 -0.076801559  0.023092384
#> on    0.019804032 -0.077998111 -0.04874882 -0.015310093 -0.045092892
#> with -0.056990459 -0.050101608 -0.12149632  0.004031857  0.007437163
#> said  0.148920147 -0.028104798  0.01654663  0.047206951  0.041610012
#> was   0.100804565  0.023544715 -0.04566950 -0.047089429 -0.059630876
#> the   0.025046480 -0.054648052 -0.06694326 -0.112027853 -0.021289741
#> at   -0.018416877 -0.040074419 -0.04154811 -0.040074419  0.057165459
#>            [,174]       [,175]       [,176]        [,177]       [,178]
#> in   -0.098557168 -0.105176651 -0.015445712 -0.0204105125  0.024639480
#> for  -0.043927293 -0.037047463 -0.001708938  0.0479846660 -0.059628567
#> that -0.070522569  0.034494393  0.006228536  0.0099169909 -0.085086432
#> is   -0.036584772 -0.018941207 -0.006227297  0.0072650913 -0.035027815
#> on   -0.053928518 -0.048139186  0.027116511 -0.0001173092 -0.007122163
#> with -0.037419889 -0.076718148 -0.062000941 -0.0040510958  0.047909643
#> said -0.058886718 -0.011618969 -0.010828590  0.1060936694  0.039906660
#> was  -0.095598791 -0.085660231 -0.053951630  0.0539516297 -0.028513995
#> the  -0.153013425 -0.003372484 -0.055785892  0.0660329846 -0.071952553
#> at   -0.001574469 -0.104312012 -0.121991441 -0.0083611582  0.062174309
#>           [,179]       [,180]      [,181]        [,182]       [,183]
#> in    0.07906626 -0.001758618 -0.01700851  0.0003795903 -0.083479749
#> for  -0.02893199  0.069507168 -0.11149429 -0.1107883171  0.020463900
#> that -0.01456465 -0.082020438 -0.11958161 -0.0620906930 -0.022613471
#> is    0.06330970  0.016605771 -0.02464934 -0.0010457650  0.003908334
#> on   -0.06185001  0.016376483  0.03336260  0.0283351537 -0.044483259
#> with -0.05072816  0.067637331  0.01996240 -0.0095117703  0.044152263
#> said -0.02482019  0.099280264  0.02774001 -0.0260366566  0.019710137
#> was   0.11026965  0.019403648  0.06034084  0.0714623569  0.008814245
#> the   0.04827241 -0.040075300 -0.04508460  0.0280067301 -0.061478825
#> at    0.05186151  0.034475973 -0.02195264  0.0633528971 -0.056281367
#>            [,184]      [,185]       [,186]        [,187]       [,188]
#> in    0.063988839 -0.09782209 -0.013147081 -0.0002703828  0.081273004
#> for   0.009482619  0.02161066 -0.008776643 -0.0698605172  0.017905912
#> that -0.014660412  0.07627150 -0.006563707 -0.0277870412  0.005916913
#> is   -0.030098160  0.01478950  0.043849864 -0.0071678479  0.038141198
#> on    0.086529232  0.02559336  0.006969287 -0.0095594485  0.020565918
#> with  0.021136626 -0.07891011 -0.034444896 -0.0058711069 -0.036636861
#> said  0.030173438  0.06229342  0.006752607 -0.0345534131  0.073486801
#> was   0.019640142  0.08234738  0.055135069 -0.0108850204  0.032655061
#> the  -0.029600639 -0.04553973 -0.032560890  0.0548756197 -0.013661547
#> at    0.039779923 -0.02313123 -0.038011738 -0.0014362726  0.067183762
#>          [,189]       [,190]        [,191]      [,192]       [,193]      [,194]
#> in   0.06693066  0.033649479  0.0189388467  0.01792811  0.061047014  0.01783623
#> for  0.13972031  0.009393739  0.0174651293 -0.02593322  0.071271746 -0.06986052
#> that 0.04560921  0.064389796  0.0226134709  0.04005180  0.002491414 -0.01456465
#> is   0.07680156  0.067980042  0.0224435633  0.07524513 -0.009599996  0.01148110
#> on   0.11395025  0.010892280 -0.0199562846  0.06733421  0.020565918 -0.02193681
#> with 0.04477817 -0.041333747 -0.0007778976 -0.03209645 -0.039454736 -0.03914178
#> said 0.01873686  0.074460074 -0.0335801407 -0.07835366 -0.092953246  0.06083326
#> was  0.04164716  0.032655061  0.0099976836  0.02318997  0.008932976  0.09938463
#> the  0.16849738  0.053054142  0.0489551188  0.05396441  0.108840034  0.04849998
#> at   0.06895195 -0.045673468 -0.0218053937  0.00163904 -0.064236989  0.04243160
#>            [,195]       [,196]       [,197]        [,198]       [,199]
#> in   -0.082743916  0.004045198 -0.013330851 -0.0255590831 -0.024823250
#> for  -0.144659974 -0.009967480  0.062098398 -0.0578639894 -0.127724507
#> that  0.011738063 -0.057107864 -0.026829409  0.0344943933 -0.038327278
#> is   -0.048779519  0.071093425 -0.103785805 -0.0267249297 -0.020108659
#> on   -0.135887061  0.063678282 -0.018280885 -0.0005902897 -0.004455876
#> with -0.003914499 -0.008219550 -0.085798964  0.0049514818 -0.091435355
#> said -0.052803392 -0.003285106  0.090033429  0.0552368222 -0.085653454
#> was  -0.022716501  0.097018239 -0.007572167 -0.0681495035  0.005324021
#> the  -0.049866324  0.017419218 -0.014572752  0.0053795604 -0.068309598
#> at   -0.163834044  0.006224854 -0.007734750  0.0424315955 -0.006703410
#>            [,200]      [,201]      [,202]        [,203]        [,204]
#> in   -0.123564188 0.072079235 -0.01379103  0.0039992554 -0.0259266229
#> for  -0.126313278 0.003704746 -0.02522725 -0.0395172938  0.0677433129
#> that -0.126480489 0.020122057  0.02874546 -0.0001208815 -0.0009882453
#> is   -0.051114424 0.054746969 -0.05682309 -0.0783585163  0.0529306966
#> on   -0.016452609 0.038542298 -0.01058778 -0.0135585676 -0.0743421833
#> with -0.116486480 0.022389086  0.04070784 -0.0205100753 -0.0601219298
#> said -0.005018856 0.044286635 -0.08565345  0.0321199829 -0.0256718663
#> was  -0.084240298 0.023544715 -0.07666813 -0.0506387763  0.0176289741
#> the  -0.108384898 0.037797754  0.05829101 -0.0403028681  0.0098479158
#> at   -0.044494276 0.047441049 -0.05569177 -0.0130386997 -0.0294665199
#>            [,205]       [,206]        [,207]      [,208]       [,209]
#> in   -0.033281186 -0.050014040 -0.0135146212 -0.02234085 -0.005723228
#> for  -0.067390686 -0.008644408 -0.0004082664  0.07056577  0.017905912
#> that -0.031236873 -0.025296412 -0.0123605237  0.04771757  0.076271503
#> is   -0.122467696  0.120910739 -0.0399574710  0.03450865  0.048779519
#> on    0.005674519  0.064287291 -0.0018756986  0.09323208  0.066115567
#> with  0.019805927 -0.043212757 -0.0302174423 -0.03100047  0.074839137
#> said -0.033093255  0.041610012  0.0299302445  0.01995333 -0.024333305
#> was   0.014611738  0.061997269 -0.0553715625  0.06862298  0.113582506
#> the   0.020265684  0.039619231  0.0305118444  0.04121314  0.053736847
#> at    0.018858621 -0.079559845 -0.0005340761  0.12022386 -0.067183762
#>            [,210]        [,211]       [,212]       [,213]      [,214]
#> in   -0.038614279 -0.0408202718  0.067298204 -0.054059238 0.011492399
#> for  -0.028755681  0.0070568661  0.085384755  0.018611888 0.088913188
#> that -0.011785944 -0.0264455707 -0.012025352  0.003664907 0.025870991
#> is    0.024908657  0.0334708582  0.037103404  0.024519683 0.051633587
#> on   -0.025135983 -0.0115019134  0.022394194 -0.023765088 0.035495380
#> with  0.008728742 -0.0186317059 -0.076718148  0.024424574 0.059808975
#> said  0.001794549 -0.0131399250 -0.009186037  0.007056599 0.005748935
#> was  -0.043540082 -0.0004991573 -0.072882290  0.087080163 0.065783110
#> the   0.024363775 -0.0170778659 -0.025161196 -0.006289833 0.004752815
#> at    0.065415578  0.0269620950 -0.049209233 -0.053923587 0.018858621
#>            [,215]       [,216]     [,217]       [,218]       [,219]
#> in   -0.062150387 -0.023903647 0.02684623 -0.015997022 -0.044681700
#> for   0.046573437  0.051160472 0.02116988 -0.035812548 -0.056452761
#> that -0.064389796  0.083553435 0.12111461  0.006898879 -0.094285198
#> is   -0.107418350  0.105861925 0.09911546 -0.063569018 -0.075763765
#> on    0.009559448  0.012339301 0.11273161  0.005103572 -0.094450720
#> with -0.003444425  0.001653914 0.02818195  0.127759261 -0.092687815
#> said  0.022629954 -0.031389904 0.05012677  0.037230037 -0.006174524
#> was  -0.100804565 -0.006181796 0.08518676 -0.067676031 -0.098438172
#> the  -0.108384898  0.003401397 0.05373685 -0.055558324 -0.082427214
#> at   -0.081917022  0.010902395 0.01370192 -0.028287932 -0.057754451
#>            [,220]       [,221]       [,222]        [,223]       [,224]
#> in   -0.009837716  0.035304161  0.017376049  0.0158132517 -0.059207809
#> for   0.020905406  0.032989368 -0.031049199  0.0187000447 -0.037400089
#> that  0.048100624  0.045993048  0.030469982 -0.0125520502 -0.034494393
#> is    0.079915473 -0.017902881 -0.077320722 -0.0003666554 -0.038919677
#> on    0.020109161  0.098107272  0.031686577 -0.0180525067  0.027421016
#> with  0.007358924  0.001546176  0.064818815  0.0541725843  0.006105823
#> said  0.014600083  0.101713694 -0.061806536 -0.0642399657 -0.030295035
#> was  -0.014552615 -0.013487907 -0.003150021  0.0615237967 -0.072408817
#> the   0.012637490  0.042351914 -0.043263119 -0.0165080130 -0.058291006
#> at    0.066594769  0.037422746  0.109615961  0.0545131825 -0.095471693
#>            [,225]       [,226]       [,227]      [,228]       [,229]
#> in   -0.006068173  0.014709879 -0.004183025  0.03125896  0.020961822
#> for   0.101614969  0.003086927 -0.027344453  0.01984680  0.043398353
#> that  0.094285198 -0.059023914  0.098118083  0.02759551  0.057107864
#> is    0.071093425  0.018941207 -0.012194747  0.01472467 -0.042033591
#> on   -0.036561770  0.001933105 -0.007883425  0.10054456  0.066115567
#> with -0.034288418 -0.012525242 -0.025207602 -0.03569736  0.118991400
#> said  0.094900289 -0.077866776 -0.002403031  0.11534001  0.009246835
#> was   0.070988884 -0.004081943  0.054425102  0.08045446 -0.007631291
#> the   0.032105754 -0.022542298  0.028804151  0.08925799  0.074230099
#> at    0.061879812 -0.064532089  0.015985472  0.01222884  0.006261666
#>           [,230]       [,231]      [,232]        [,233]        [,234]
#> in    0.01025120  0.026110393 -0.13753899  0.0904667722  5.589844e-02
#> for   0.02046390  0.020287587 -0.02646216  0.0945581033 -7.009175e-05
#> that  0.06860573  0.016576461 -0.05787397  0.0270209352 -7.397162e-02
#> is    0.11001310 -0.044368495 -0.02633596  0.0167354291  7.887768e-02
#> on   -0.02117555  0.113341239 -0.02787840  0.0213278034 -2.361284e-02
#> with  0.01088159  0.059808975 -0.03100047  0.0162832632  1.503080e-02
#> said  0.03333695  0.001383911  0.02360323 -0.0540198579 -7.397369e-02
#> was   0.05963088  0.035021454 -0.02117832 -0.0366778810 -4.377658e-02
#> the   0.03665991  0.026071469 -0.08015060  0.0755964403  6.193396e-02
#> at   -0.05745996  0.026225251 -0.12670640 -0.0005226101 -6.445726e-03
#>             [,235]       [,236]       [,237]       [,238]      [,239]
#> in   -0.0308914236 -0.007493143  0.032362336 -0.005493516  0.09267352
#> for  -0.0501018703 -0.015965744  0.049043268 -0.016847310 -0.01107016
#> that  0.0091030033 -0.044843104 -0.061706855  0.012552050  0.05940775
#> is    0.0295789968 -0.023870863 -0.042292907  0.002529922 -0.01102729
#> on   -0.0055222662 -0.023003203  0.049053324  0.016909990  0.02879253
#> with -0.0191793766 -0.025989989  0.005675510  0.059182424  0.01557847
#> said  0.0281047981 -0.035039800  0.009307634 -0.047206951  0.02031812
#> was   0.0492188435 -0.070042424 -0.019403648 -0.006152234  0.01289619
#> the  -0.0003917156 -0.064666643  0.033472095 -0.031878186  0.04189678
#> at   -0.0034072248 -0.044494276  0.044789376  0.061879812  0.07484489
#>            [,240]       [,241]       [,242]       [,243]      [,244]     [,245]
#> in    0.043394557 -0.040268962 -0.024271940 -0.006849196 -0.03512039 0.03309742
#> for  -0.042163438  0.044103606  0.000466074  0.002888936 -0.05151310 0.06633208
#> that  0.023954941  0.003617025 -0.114215728 -0.019450929 -0.08470338 0.05404187
#> is    0.031914432  0.002529922  0.005935035  0.091850905 -0.07161259 0.01634646
#> on    0.042960111  0.031534324 -0.001090102 -0.085310590 -0.09566936 0.06154550
#> with  0.015500235 -0.103334687  0.101455676  0.021136626 -0.03961185 0.04884915
#> said -0.021169796  0.029808149  0.018980058 -0.068619941  0.01691192 0.07981332
#> was  -0.037860835  0.017274233  0.016919008  0.087080163 -0.06720304 0.01360615
#> the  -0.007201038 -0.006916578 -0.044400959  0.013035734 -0.09290094 0.02299743
#> at    0.093114517  0.002689088  0.067183762 -0.071898719 -0.13967147 0.01480688
#>            [,246]       [,247]       [,248]       [,249]      [,250]
#> in   -0.038245986  0.051853246  0.002252688 -0.003148943 -0.03328119
#> for   0.018964515  0.014466359  0.025579875 -0.041810089 -0.02143435
#> that  0.045609210  0.098118083 -0.051358929  0.004144508  0.00929453
#> is   -0.042292907  0.048001041  0.032433064  0.040995265 -0.02672493
#> on   -0.100544557  0.004132028 -0.053624013 -0.010968406  0.02544049
#> with  0.056990459  0.129011721  0.045404723 -0.050728159  0.03585383
#> said  0.049153496 -0.019710137 -0.064239966  0.094900289 -0.01399160
#> was  -0.013902256  0.032891555  0.034074509  0.046379463  0.02425468
#> the  -0.092900942  0.107018557  0.029600639  0.020607037  0.06739933
#> at    0.003591285  0.011786487 -0.019595465 -0.090757340 -0.06954094
#>            [,251]       [,252]      [,253]       [,254]       [,255]
#> in    0.055530151 -0.009607251  0.05074987  0.004735088  0.056633523
#> for   0.019758647  0.018170382  0.04357467  0.095264079 -0.003153406
#> that  0.054807976  0.099651080  0.05135893 -0.013606231  0.093519092
#> is   -0.049038835 -0.142186850  0.05734172  0.045666137  0.033470858
#> on    0.018661516  0.007122163 -0.01264443 -0.040217698  0.010892280
#> with  0.002798636  0.009550890  0.11335501 -0.004912362 -0.018475229
#> said -0.052560198  0.005231651  0.04136682 -0.015816549  0.090033429
#> was  -0.010707650 -0.145764095  0.05537156  0.036440903  0.006122673
#> the   0.034382367  0.023908639  0.01275127 -0.025502548  0.005521324
#> at   -0.088399560 -0.005451198  0.01723768 -0.080148838  0.091935928
#>            [,256]        [,257]       [,258]       [,259]       [,260]
#> in   -0.028500908  0.0036776580  0.033649479 -0.050749873  0.007309374
#> for   0.001973408  0.0432213175  0.071271746 -0.066332084 -0.033165681
#> that -0.025487938  0.1134488375  0.060173858 -0.025296412 -0.051741982
#> is    0.057341721 -0.0142055066  0.054228338 -0.064087649  0.028152229
#> on   -0.080435396 -0.0002165228  0.026354626  0.002056654  0.076474964
#> with -0.042899803  0.0522935736 -0.006888851 -0.053546034  0.067950286
#> said -0.057183367  0.0307819202  0.001079919 -0.017641497 -0.099766651
#> was   0.064363178  0.0122458297  0.017628974 -0.013665278 -0.066256583
#> the  -0.062845166  0.0471336414 -0.026412821 -0.042124345 -0.016166661
#> at   -0.035801810  0.0307923564  0.068362350 -0.069836038  0.033886377
#>            [,261]      [,262]      [,263]      [,264]      [,265]     [,266]
#> in    0.003562425  0.01544571  0.05369170  0.12871276  0.13091950 0.04192364
#> for  -0.012613624  0.02769708 -0.01384854  0.03351903  0.03457763 0.07091912
#> that  0.049442095 -0.04905904 -0.07550461  0.08355343 -0.03123687 0.09121920
#> is    0.050336477 -0.08769920  0.02348136  0.03840105  0.02192493 0.22625350
#> on   -0.059717226  0.05758445  0.06550593 -0.07769361  0.02239419 0.09079479
#> with  0.024894006  0.02458105 -0.02019712  0.03162638  0.01526552 0.05010161
#> said -0.039663467 -0.04939669  0.01478248  0.02153508  0.05012677 0.15184046
#> was   0.087080163 -0.04495953  0.04117369  0.06436318  0.01928540 0.20823484
#> the   0.019695832  0.03278846 -0.04007530  0.06193396  0.11384933 0.11521567
#> at    0.048914737  0.01009253  0.07013053  0.01679594  0.08898916 0.01038703
#>            [,267]      [,268]       [,269]       [,270]      [,271]
#> in    0.068769870 -0.02813337  0.037510907 -0.029604281  0.03328119
#> for  -0.029108308  0.06809594 -0.025050935 -0.030519537  0.05080712
#> that -0.007617888 -0.02778704 -0.051358929  0.046184574  0.12724659
#> is   -0.054746969 -0.08510444 -0.048001041 -0.033990021 -0.02555748
#> on   -0.065811062  0.01431983 -0.005217762  0.002894666  0.13710633
#> with  0.007945714 -0.02066719  0.041960297  0.037419889  0.08830388
#> said  0.134320064 -0.04525991 -0.044286635  0.096846834  0.11728705
#> was  -0.009110347 -0.07146236  0.029578702  0.043776576  0.04212015
#> the   0.003728759  0.04212435 -0.017305434  0.045084596  0.04212435
#> at    0.054513183  0.04596796  0.087220971 -0.014806884 -0.03138195
#>            [,272]      [,273]       [,274]      [,275]       [,276]
#> in    0.047072215  0.03659130 -0.040085192  0.03677507 -0.098557168
#> for  -0.009879323  0.07691666  0.011907649  0.09526408 -0.001224077
#> that  0.040244113  0.12494749  0.074738506  0.05979081 -0.072055565
#> is    0.076801559  0.03476850  0.026205767  0.02880052  0.072131219
#> on   -0.030925315  0.14868374 -0.036409518  0.03016343  0.037780413
#> with -0.020510075 -0.03616679 -0.046030632  0.05479849  0.018083394
#> said -0.103660239  0.01606024  0.009368432 -0.06910683 -0.012957530
#> was  -0.033365028  0.05371514  0.047562901  0.07288229  0.078088064
#> the   0.081060872  0.02743781  0.035066004  0.03210575 -0.068764734
#> at    0.028435180  0.04361079  0.055397275  0.07248771 -0.085453390
#>            [,277]      [,278]      [,279]      [,280]       [,281]       [,282]
#> in   -0.021789540 -0.02721377 -0.04578507 -0.04321079  0.092673518 -0.062150387
#> for  -0.006130499 -0.10302620 -0.03369534 -0.07938649  0.059275218 -0.029637970
#> that  0.019738219 -0.06170686  0.07013873 -0.04599305 -0.031428399  0.036027390
#> is   -0.010248816 -0.11468344 -0.03943884 -0.07783935  0.006097639 -0.087699196
#> on   -0.023308331 -0.11090333  0.02803065 -0.02635463  0.051491233  0.068857980
#> with  0.001291579 -0.04540472  0.02223261 -0.02567703 -0.020040643  0.057303414
#> said  0.007847476 -0.01429609  0.03917658 -0.05742656 -0.090520314  0.082246751
#> was  -0.039043789 -0.08471377 -0.01609079 -0.06247026  0.059394382 -0.007010009
#> the  -0.037570186 -0.13661920 -0.02276987 -0.01821571  0.006148069 -0.001693703
#> at   -0.013775544 -0.07513999 -0.04066401 -0.03565456 -0.022689486 -0.048619637
#>            [,283]       [,284]       [,285]       [,286]       [,287]
#> in   -0.008964056  0.094144430  0.001022785  0.048175588 -0.080169631
#> for  -0.013672226  0.063509627 -0.002029048  0.172180740 -0.034047970
#> that  0.024337994  0.030661509  0.027979353 -0.083170382 -0.029128512
#> is   -0.055784764  0.001702557  0.071612588 -0.002108003 -0.055006817
#> on   -0.068248347  0.058803088 -0.045092892  0.030620186 -0.098716281
#> with -0.025833511  0.029434414  0.070455206  0.051667023 -0.056364550
#> said -0.057670252 -0.058400331  0.085166568  0.062780306 -0.006235322
#> was  -0.079034524 -0.017392480  0.013487907 -0.030762141 -0.038334308
#> the  -0.010189268  0.086981373  0.061023689  0.017191650 -0.086981373
#> at   -0.041842603  0.031676449  0.049503729  0.011712863 -0.048914737
#>           [,288]       [,289]        [,290]       [,291]       [,292]
#> in   -0.10811923 -0.031810273  0.0181118821 -0.127241846 -0.066930665
#> for  -0.01658284  0.029460935  0.0211698756 -0.016318371  0.002690222
#> that -0.12648049  0.016767988  0.0009584174 -0.008863595 -0.012264760
#> is   -0.07420681  0.055525448 -0.0066810996 -0.124024121 -0.019330180
#> on    0.04874882  0.018433138 -0.0755608258  0.016757114 -0.017442874
#> with -0.02129310  0.031469902  0.0089634583 -0.013308270  0.001976488
#> said -0.01642503  0.026036657  0.0135660120 -0.096359949  0.014904075
#> was  -0.05560806  0.008932976  0.0141387503 -0.145764095 -0.022243029
#> the  -0.01468654 -0.066488121 -0.0833374859 -0.066488121 -0.028120514
#> at   -0.05628137  0.041253007 -0.0128178275 -0.019889961  0.031823697
#>           [,293]      [,294]       [,295]       [,296]       [,297]
#> in   -0.06067872  0.04891067  0.046152612 -0.035671701 -0.044314160
#> for  -0.05962857  0.05892259  0.005733794  0.000344678  0.013319600
#> that -0.02625404 -0.01619341 -0.015234991  0.050208986  0.015810355
#> is   -0.04981731  0.09703988  0.014400525  0.067980042 -0.013167712
#> on    0.01927115  0.02528824 -0.081654038  0.051795737  0.009825890
#> with -0.08893043 -0.01565671 -0.059808975 -0.030530397 -0.048536193
#> said -0.04647687  0.02676674 -0.058643525  0.034796607  0.052803392
#> was  -0.08471377  0.11310903  0.010944144  0.053478642 -0.050165788
#> the  -0.01212453  0.01525546 -0.017077866  0.013833155  0.004667944
#> at   -0.06895195  0.11315173  0.001786289 -0.001657748 -0.028435180
#>            [,298]      [,299]      [,300]
#> in   -0.035855471  0.01089439 -0.04707222
#> for   0.051513099 -0.02522725  0.01746513
#> that  0.005389430  0.04790910 -0.11651562
#> is    0.005967450  0.08718003  0.05682309
#> on   -0.072818412 -0.01835701 -0.04417875
#> with  0.006927970 -0.01205581 -0.04415226
#> said  0.072026643  0.09003343 -0.04331336
#> was  -0.058920910  0.10743027 -0.01064853
#> the   0.003415387  0.04440096 -0.06421151
#> at   -0.065121081 -0.05215601 -0.01930037

d.dense = data_wordvec_reshape(d.plain, to="dense")
d.dense  # identical to `d`
#>     word
#>  1:   in
#>  2:  for
#>  3: that
#>  4:   is
#>  5:   on
#>  6: with
#>  7: said
#>  8:  was
#>  9:  the
#> 10:   at
#>                                                                                         vec
#>  1:              0.05295587, 0.06545975, 0.06619558, 0.04707222, 0.05222079,-0.08200884,...
#>  2:       -0.008512173,-0.034224283, 0.032284114, 0.045868184,-0.013142564,-0.046220811,...
#>  3:             -0.01236052,-0.02222963, 0.06553974, 0.03947722,-0.08662021, 0.02491257,...
#>  4:        0.003746262,-0.038919677, 0.091331742, 0.012000260,-0.070574262, 0.105342762,...
#>  5:              0.01668099,-0.05667031, 0.01736675, 0.12735656, 0.00388493,-0.05636580,...
#>  6:             -0.01596967, 0.01409130,-0.02270204, 0.08767797, 0.01048975,-0.01792692,...
#>  7:       -0.004531971,-0.022021472, 0.049639883,-0.037960116,-0.028226893, 0.030659825,...
#>  8:  0.0126005706,-0.0009168986, 0.0899195443,-0.0250828942, 0.0024846400,-0.0532416633,...
#>  9:              0.07468524, 0.09791024, 0.04645094, 0.04986632,-0.06284517,-0.11248299,...
#> 10:             -0.03536007,-0.02268949, 0.04390528, 0.06571068, 0.04007442, 0.02931927,...