Normalize all word vectors to the unit length 1.

L2-normalization (scaling to unit euclidean length): the norm of each vector in the vector space will be normalized to 1. It is necessary for any linear operation of word vectors.

R code:

Vector: vec / sqrt(sum(vec^2))

Matrix: mat / sqrt(rowSums(mat^2))

Usage

normalize(x)

Arguments

x

A wordvec (data.table) or embed (matrix), see data_wordvec_load.

Value

A wordvec (data.table) or embed (matrix) with normalized word vectors.

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

See also

Examples

d = normalize(demodata)
# the same: d = as_wordvec(demodata, normalize=TRUE)