Tabulate data for Word Embedding Association Test (WEAT).

Tabulate data (cosine similarity and standardized effect size) for Word Embedding Association Test (WEAT) and Single-Category Word Embedding Association Test (SC-WEAT) analyses.

Usage

tab_WEAT(data, T1, T2, A1, A2, use.pattern = FALSE, labels)

Arguments

data

A data.table with variables word and vec loaded by data_wordvec_load.

T1, T2

Target words (a vector of words or a pattern of regular expression). If only T1 is specified, it will tabulate data for single-category WEAT (SC-WEAT).

A1, A2

Attribute words (a vector of words or a pattern of regular expression). Both must be specified.

use.pattern

Defaults to FALSE (using a vector of words). If you use regular expression in T1, T2, A1, and A2, please specify this argument as TRUE.

labels

Labels for target and attribute concepts (a named list), such as (the default) list(T1="Target1", T2="Target2", A1="Attrib1", A2="Attrib2").

Value

A list of objects:

words.valid

valid (actually matched) words

data.raw

data.table of cosine similarities between all word pairs

data.mean

data.table of mean cosine similarities across all attribute words

data.diff

data.table of differential mean cosine similarities between the two attribute concepts

code.diff

description for the difference between the two attribute concepts

eff.type

effect type: WEAT or SC-WEAT

eff.raw

raw effect for WEAT (a single value) or SC-WEAT (a data table)

eff.size

standardized effect size for WEAT (a single value) or SC-WEAT (a data table)

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

Caliskan, A., Bryson, J. J., & Narayanan, A. (2017). Semantics derived automatically from language corpora contain human-like biases. Science, 356(6334), 183--186.

See also

Examples

## Remember: cc() is more convenient than c()!

weat = tab_WEAT(
  demodata,
  T1=cc("king, King"),
  T2=cc("queen, Queen"),
  A1=cc("male, man, boy, brother, he, him, his, son"),
  A2=cc("female, woman, girl, sister, she, her, hers, daughter"),
  labels=list(T1="King", T2="Queen", A1="Male", A2="Female"))
weat
#> $words.valid
#> $words.valid$T1
#> [1] "king" "King"
#> 
#> $words.valid$T2
#> [1] "queen" "Queen"
#> 
#> $words.valid$A1
#> [1] "male"    "man"     "boy"     "brother" "he"      "him"     "his"    
#> [8] "son"    
#> 
#> $words.valid$A2
#> [1] "female"   "woman"    "girl"     "sister"   "she"      "her"      "hers"    
#> [8] "daughter"
#> 
#> 
#> $data.raw
#>     Target Attrib T_word   A_word    cos_sim
#>  1:   King   Male   king     male 0.13960424
#>  2:   King   Male   King     male 0.09070985
#>  3:   King   Male   king      man 0.22942676
#>  4:   King   Male   King      man 0.15777646
#>  5:   King   Male   king      boy 0.25351544
#>  6:   King   Male   King      boy 0.16933196
#>  7:   King   Male   king  brother 0.23432925
#>  8:   King   Male   King  brother 0.22089012
#>  9:   King   Male   king       he 0.22512579
#> 10:   King   Male   King       he 0.14808663
#> 11:   King   Male   king      him 0.24867882
#> 12:   King   Male   King      him 0.21191115
#> 13:   King   Male   king      his 0.17143112
#> 14:   King   Male   King      his 0.13784300
#> 15:   King   Male   king      son 0.21417183
#> 16:   King   Male   King      son 0.18987593
#> 17:   King Female   king   female 0.08971824
#> 18:   King Female   King   female 0.07212608
#> 19:   King Female   king    woman 0.12847968
#> 20:   King Female   King    woman 0.06369529
#> 21:   King Female   king     girl 0.17236953
#> 22:   King Female   King     girl 0.08804782
#> 23:   King Female   king   sister 0.11340389
#> 24:   King Female   King   sister 0.08357969
#> 25:   King Female   king      she 0.06542464
#> 26:   King Female   King      she 0.04830619
#> 27:   King Female   king      her 0.06532056
#> 28:   King Female   King      her 0.07439990
#> 29:   King Female   king     hers 0.10306190
#> 30:   King Female   King     hers 0.06238802
#> 31:   King Female   king daughter 0.14881282
#> 32:   King Female   King daughter 0.11301357
#> 33:  Queen   Male  queen     male 0.18550913
#> 34:  Queen   Male  Queen     male 0.11256271
#> 35:  Queen   Male  queen      man 0.16658216
#> 36:  Queen   Male  Queen      man 0.09365463
#> 37:  Queen   Male  queen      boy 0.20767608
#> 38:  Queen   Male  Queen      boy 0.10297925
#> 39:  Queen   Male  queen  brother 0.11763060
#> 40:  Queen   Male  Queen  brother 0.04219659
#> 41:  Queen   Male  queen       he 0.09282878
#> 42:  Queen   Male  Queen       he 0.05188371
#> 43:  Queen   Male  queen      him 0.13145000
#> 44:  Queen   Male  Queen      him 0.07582127
#> 45:  Queen   Male  queen      his 0.06322216
#> 46:  Queen   Male  Queen      his 0.02713799
#> 47:  Queen   Male  queen      son 0.15659811
#> 48:  Queen   Male  Queen      son 0.05476228
#> 49:  Queen Female  queen   female 0.26985589
#> 50:  Queen Female  Queen   female 0.18863183
#> 51:  Queen Female  queen    woman 0.31618132
#> 52:  Queen Female  Queen    woman 0.20171619
#> 53:  Queen Female  queen     girl 0.35005602
#> 54:  Queen Female  Queen     girl 0.18701451
#> 55:  Queen Female  queen   sister 0.25180591
#> 56:  Queen Female  Queen   sister 0.16349528
#> 57:  Queen Female  queen      she 0.36242807
#> 58:  Queen Female  Queen      she 0.23824302
#> 59:  Queen Female  queen      her 0.35522259
#> 60:  Queen Female  Queen      her 0.23308440
#> 61:  Queen Female  queen     hers 0.31618069
#> 62:  Queen Female  Queen     hers 0.21118412
#> 63:  Queen Female  queen daughter 0.30513862
#> 64:  Queen Female  Queen daughter 0.16699306
#>     Target Attrib T_word   A_word    cos_sim
#> 
#> $data.mean
#>    Target T_word Attrib cos_sim_mean    std_dev  std_mean
#> 1:   King   king   Male   0.21453541 0.06536807 3.2819604
#> 2:   King   king Female   0.11082391 0.06536807 1.6953830
#> 3:   King   King   Male   0.16580314 0.05640498 2.9395122
#> 4:   King   King Female   0.07569457 0.05640498 1.3419837
#> 5:  Queen  queen   Male   0.14018713 0.10032951 1.3972671
#> 6:  Queen  queen Female   0.31585864 0.10032951 3.1482127
#> 7:  Queen  Queen   Male   0.07012480 0.07224076 0.9707096
#> 8:  Queen  Queen Female   0.19879530 0.07224076 2.7518439
#> 
#> $data.diff
#>    Target T_word cos_sim_diff   std_dev
#> 1:   King   king   0.10371150 0.1451877
#> 2:   King   King   0.09010857 0.1451877
#> 3:  Queen  queen  -0.17567151 0.1451877
#> 4:  Queen  Queen  -0.12867050 0.1451877
#> 
#> $code.diff
#> [1] "King vs. Queen :: Male vs. Female"
#> 
#> $eff.type
#> [1] "WEAT (Word Embedding Association Test)"
#> 
#> $eff.raw
#> [1] 0.249081
#> 
#> $eff.size
#> [1] 1.715579
#> 

weat = tab_WEAT(
  demodata,
  T1="^[kK]ing$",
  T2="^[qQ]ueen$",
  A1="^male$|^man$|^boy$|^brother$|^he$|^him$|^his$|^son$",
  A2="^female$|^woman$|^girl$|^sister$|^she$|^her$|^hers$|^daughter$",
  use.pattern=TRUE,
  labels=list(T1="King", T2="Queen", A1="Male", A2="Female"))
#> T1 (King):
#> 2 words are matched...
#> T2 (Queen):
#> 2 words are matched...
#> A1 (Male):
#> 8 words are matched...
#> A2 (Female):
#> 8 words are matched...
weat
#> $words.valid
#> $words.valid$T1
#> [1] "King" "king"
#> 
#> $words.valid$T2
#> [1] "Queen" "queen"
#> 
#> $words.valid$A1
#> [1] "he"      "his"     "him"     "man"     "son"     "boy"     "brother"
#> [8] "male"   
#> 
#> $words.valid$A2
#> [1] "her"      "she"      "woman"    "daughter" "girl"     "female"   "sister"  
#> [8] "hers"    
#> 
#> 
#> $data.raw
#>     Target Attrib T_word   A_word    cos_sim
#>  1:   King   Male   King       he 0.14808663
#>  2:   King   Male   king       he 0.22512579
#>  3:   King   Male   King      his 0.13784300
#>  4:   King   Male   king      his 0.17143112
#>  5:   King   Male   King      him 0.21191115
#>  6:   King   Male   king      him 0.24867882
#>  7:   King   Male   King      man 0.15777646
#>  8:   King   Male   king      man 0.22942676
#>  9:   King   Male   King      son 0.18987593
#> 10:   King   Male   king      son 0.21417183
#> 11:   King   Male   King      boy 0.16933196
#> 12:   King   Male   king      boy 0.25351544
#> 13:   King   Male   King  brother 0.22089012
#> 14:   King   Male   king  brother 0.23432925
#> 15:   King   Male   King     male 0.09070985
#> 16:   King   Male   king     male 0.13960424
#> 17:   King Female   King      her 0.07439990
#> 18:   King Female   king      her 0.06532056
#> 19:   King Female   King      she 0.04830619
#> 20:   King Female   king      she 0.06542464
#> 21:   King Female   King    woman 0.06369529
#> 22:   King Female   king    woman 0.12847968
#> 23:   King Female   King daughter 0.11301357
#> 24:   King Female   king daughter 0.14881282
#> 25:   King Female   King     girl 0.08804782
#> 26:   King Female   king     girl 0.17236953
#> 27:   King Female   King   female 0.07212608
#> 28:   King Female   king   female 0.08971824
#> 29:   King Female   King   sister 0.08357969
#> 30:   King Female   king   sister 0.11340389
#> 31:   King Female   King     hers 0.06238802
#> 32:   King Female   king     hers 0.10306190
#> 33:  Queen   Male  Queen       he 0.05188371
#> 34:  Queen   Male  queen       he 0.09282878
#> 35:  Queen   Male  Queen      his 0.02713799
#> 36:  Queen   Male  queen      his 0.06322216
#> 37:  Queen   Male  Queen      him 0.07582127
#> 38:  Queen   Male  queen      him 0.13145000
#> 39:  Queen   Male  Queen      man 0.09365463
#> 40:  Queen   Male  queen      man 0.16658216
#> 41:  Queen   Male  Queen      son 0.05476228
#> 42:  Queen   Male  queen      son 0.15659811
#> 43:  Queen   Male  Queen      boy 0.10297925
#> 44:  Queen   Male  queen      boy 0.20767608
#> 45:  Queen   Male  Queen  brother 0.04219659
#> 46:  Queen   Male  queen  brother 0.11763060
#> 47:  Queen   Male  Queen     male 0.11256271
#> 48:  Queen   Male  queen     male 0.18550913
#> 49:  Queen Female  Queen      her 0.23308440
#> 50:  Queen Female  queen      her 0.35522259
#> 51:  Queen Female  Queen      she 0.23824302
#> 52:  Queen Female  queen      she 0.36242807
#> 53:  Queen Female  Queen    woman 0.20171619
#> 54:  Queen Female  queen    woman 0.31618132
#> 55:  Queen Female  Queen daughter 0.16699306
#> 56:  Queen Female  queen daughter 0.30513862
#> 57:  Queen Female  Queen     girl 0.18701451
#> 58:  Queen Female  queen     girl 0.35005602
#> 59:  Queen Female  Queen   female 0.18863183
#> 60:  Queen Female  queen   female 0.26985589
#> 61:  Queen Female  Queen   sister 0.16349528
#> 62:  Queen Female  queen   sister 0.25180591
#> 63:  Queen Female  Queen     hers 0.21118412
#> 64:  Queen Female  queen     hers 0.31618069
#>     Target Attrib T_word   A_word    cos_sim
#> 
#> $data.mean
#>    Target T_word Attrib cos_sim_mean    std_dev  std_mean
#> 1:   King   King   Male   0.16580314 0.05640498 2.9395122
#> 2:   King   King Female   0.07569457 0.05640498 1.3419837
#> 3:   King   king   Male   0.21453541 0.06536807 3.2819604
#> 4:   King   king Female   0.11082391 0.06536807 1.6953830
#> 5:  Queen  Queen   Male   0.07012480 0.07224076 0.9707096
#> 6:  Queen  Queen Female   0.19879530 0.07224076 2.7518439
#> 7:  Queen  queen   Male   0.14018713 0.10032951 1.3972671
#> 8:  Queen  queen Female   0.31585864 0.10032951 3.1482127
#> 
#> $data.diff
#>    Target T_word cos_sim_diff   std_dev
#> 1:   King   King   0.09010857 0.1451877
#> 2:   King   king   0.10371150 0.1451877
#> 3:  Queen  Queen  -0.12867050 0.1451877
#> 4:  Queen  queen  -0.17567151 0.1451877
#> 
#> $code.diff
#> [1] "King vs. Queen :: Male vs. Female"
#> 
#> $eff.type
#> [1] "WEAT (Word Embedding Association Test)"
#> 
#> $eff.raw
#> [1] 0.249081
#> 
#> $eff.size
#> [1] 1.715579
#> 

sc_weat = tab_WEAT(
  demodata,
  T1=cc("
    architect, boss, leader, engineer, CEO, officer, manager,
    lawyer, scientist, doctor, psychologist, investigator,
    consultant, programmer, teacher, clerk, counselor,
    salesperson, therapist, psychotherapist, nurse"),
  A1=cc("male, man, boy, brother, he, him, his, son"),
  A2=cc("female, woman, girl, sister, she, her, hers, daughter"),
  labels=list(T1="Occupation", A1="Male", A2="Female"))
sc_weat
#> $words.valid
#> $words.valid$T1
#>  [1] "architect"       "boss"            "leader"          "engineer"       
#>  [5] "CEO"             "officer"         "manager"         "lawyer"         
#>  [9] "scientist"       "doctor"          "psychologist"    "investigator"   
#> [13] "consultant"      "programmer"      "teacher"         "clerk"          
#> [17] "counselor"       "salesperson"     "therapist"       "psychotherapist"
#> [21] "nurse"          
#> 
#> $words.valid$T2
#> NULL
#> 
#> $words.valid$A1
#> [1] "male"    "man"     "boy"     "brother" "he"      "him"     "his"    
#> [8] "son"    
#> 
#> $words.valid$A2
#> [1] "female"   "woman"    "girl"     "sister"   "she"      "her"      "hers"    
#> [8] "daughter"
#> 
#> 
#> $data.raw
#>          Target Attrib          T_word   A_word     cos_sim
#>   1: Occupation   Male       architect     male -0.07107298
#>   2: Occupation   Male            boss     male  0.02509828
#>   3: Occupation   Male          leader     male -0.04058901
#>   4: Occupation   Male        engineer     male -0.04795309
#>   5: Occupation   Male             CEO     male -0.11093326
#>  ---                                                       
#> 332: Occupation Female       counselor daughter  0.24606261
#> 333: Occupation Female     salesperson daughter  0.13918887
#> 334: Occupation Female       therapist daughter  0.29122729
#> 335: Occupation Female psychotherapist daughter  0.26725654
#> 336: Occupation Female           nurse daughter  0.36904112
#> 
#> $data.mean
#>         Target          T_word Attrib  cos_sim_mean    std_dev     std_mean
#>  1: Occupation       architect   Male  8.692525e-02 0.08302492  1.046977869
#>  2: Occupation       architect Female -8.863421e-05 0.08302492 -0.001067562
#>  3: Occupation            boss   Male  2.060022e-01 0.07892258  2.610181026
#>  4: Occupation            boss Female  1.249453e-01 0.07892258  1.583138058
#>  5: Occupation          leader   Male  1.237260e-01 0.08049820  1.537003534
#>  6: Occupation          leader Female  4.439556e-02 0.08049820  0.551509980
#>  7: Occupation        engineer   Male  1.030868e-01 0.09144904  1.127259846
#>  8: Occupation        engineer Female  3.219294e-02 0.09144904  0.352031432
#>  9: Occupation             CEO   Male -2.858738e-02 0.07028655 -0.406726174
#> 10: Occupation             CEO Female -7.334349e-02 0.07028655 -1.043492558
#> 11: Occupation         officer   Male  1.308768e-01 0.07290262  1.795228269
#> 12: Occupation         officer Female  9.821684e-02 0.07290262  1.347233392
#> 13: Occupation         manager   Male  4.827265e-02 0.06566996  0.735079676
#> 14: Occupation         manager Female  2.579163e-02 0.06566996  0.392746207
#> 15: Occupation          lawyer   Male  1.848924e-01 0.08545002  2.163749414
#> 16: Occupation          lawyer Female  1.645807e-01 0.08545002  1.926046604
#> 17: Occupation       scientist   Male  9.010405e-02 0.04787808  1.881947739
#> 18: Occupation       scientist Female  8.196547e-02 0.04787808  1.711962285
#> 19: Occupation          doctor   Male  2.683077e-01 0.07933172  3.382098104
#> 20: Occupation          doctor Female  2.549294e-01 0.07933172  3.213461245
#> 21: Occupation    psychologist   Male  1.508973e-01 0.03388385  4.453369388
#> 22: Occupation    psychologist Female  1.498728e-01 0.03388385  4.423133632
#> 23: Occupation    investigator   Male  1.343693e-01 0.05140167  2.614104205
#> 24: Occupation    investigator Female  1.337600e-01 0.05140167  2.602249123
#> 25: Occupation      consultant   Male  4.682858e-02 0.05764734  0.812328417
#> 26: Occupation      consultant Female  5.000916e-02 0.05764734  0.867501577
#> 27: Occupation      programmer   Male  4.665029e-02 0.04187294  1.114091656
#> 28: Occupation      programmer Female  5.267160e-02 0.04187294  1.257891213
#> 29: Occupation         teacher   Male  2.060090e-01 0.08399303  2.452691471
#> 30: Occupation         teacher Female  2.731964e-01 0.08399303  3.252607343
#> 31: Occupation           clerk   Male  1.870184e-01 0.07531653  2.483099547
#> 32: Occupation           clerk Female  2.649577e-01 0.07531653  3.517923009
#> 33: Occupation       counselor   Male  1.408918e-01 0.05387167  2.615322725
#> 34: Occupation       counselor Female  2.053143e-01 0.05387167  3.811174881
#> 35: Occupation     salesperson   Male  1.029099e-01 0.06846796  1.503037241
#> 36: Occupation     salesperson Female  1.855234e-01 0.06846796  2.709638701
#> 37: Occupation       therapist   Male  1.730993e-01 0.05566177  3.109841850
#> 38: Occupation       therapist Female  2.423886e-01 0.05566177  4.354669527
#> 39: Occupation psychotherapist   Male  1.119207e-01 0.06973167  1.605019706
#> 40: Occupation psychotherapist Female  2.041495e-01 0.06973167  2.927643750
#> 41: Occupation           nurse   Male  1.796695e-01 0.10881512  1.651144349
#> 42: Occupation           nurse Female  3.418709e-01 0.10881512  3.141758950
#>         Target          T_word Attrib  cos_sim_mean    std_dev     std_mean
#> 
#> $data.diff
#>         Target          T_word cos_sim_diff    std_diff
#>  1: Occupation       architect  0.087013885  1.04804543
#>  2: Occupation            boss  0.081056880  1.02704297
#>  3: Occupation          leader  0.079330460  0.98549355
#>  4: Occupation        engineer  0.070893895  0.77522841
#>  5: Occupation             CEO  0.044756111  0.63676638
#>  6: Occupation         officer  0.032659998  0.44799488
#>  7: Occupation         manager  0.022481025  0.34233347
#>  8: Occupation          lawyer  0.020311709  0.23770281
#>  9: Occupation       scientist  0.008138578  0.16998545
#> 10: Occupation          doctor  0.013378252  0.16863686
#> 11: Occupation    psychologist  0.001024504  0.03023576
#> 12: Occupation    investigator  0.000609371  0.01185508
#> 13: Occupation      consultant -0.003180586 -0.05517316
#> 14: Occupation      programmer -0.006021310 -0.14379956
#> 15: Occupation         teacher -0.067187362 -0.79991587
#> 16: Occupation           clerk -0.077939310 -1.03482346
#> 17: Occupation       counselor -0.064422548 -1.19585216
#> 18: Occupation     salesperson -0.082613539 -1.20660146
#> 19: Occupation       therapist -0.069289306 -1.24482768
#> 20: Occupation psychotherapist -0.092228780 -1.32262404
#> 21: Occupation           nurse -0.162201404 -1.49061460
#>         Target          T_word cos_sim_diff    std_diff
#> 
#> $code.diff
#> [1] "Occupation :: Male vs. Female"
#> 
#> $eff.type
#> [1] "SC-WEAT (Single-Category Word Embedding Association Test)"
#> 
#> $eff.raw
#>              T_word      eff_raw closer_to
#>  1:       architect  0.087013885      Male
#>  2:            boss  0.081056880      Male
#>  3:          leader  0.079330460      Male
#>  4:        engineer  0.070893895      Male
#>  5:             CEO  0.044756111      Male
#>  6:         officer  0.032659998      Male
#>  7:         manager  0.022481025      Male
#>  8:          lawyer  0.020311709      Male
#>  9:       scientist  0.008138578      Male
#> 10:          doctor  0.013378252      Male
#> 11:    psychologist  0.001024504      Male
#> 12:    investigator  0.000609371      Male
#> 13:      consultant -0.003180586    Female
#> 14:      programmer -0.006021310    Female
#> 15:         teacher -0.067187362    Female
#> 16:           clerk -0.077939310    Female
#> 17:       counselor -0.064422548    Female
#> 18:     salesperson -0.082613539    Female
#> 19:       therapist -0.069289306    Female
#> 20: psychotherapist -0.092228780    Female
#> 21:           nurse -0.162201404    Female
#>              T_word      eff_raw closer_to
#> 
#> $eff.size
#>              T_word    eff_size closer_to
#>  1:       architect  1.04804543      Male
#>  2:            boss  1.02704297      Male
#>  3:          leader  0.98549355      Male
#>  4:        engineer  0.77522841      Male
#>  5:             CEO  0.63676638      Male
#>  6:         officer  0.44799488      Male
#>  7:         manager  0.34233347      Male
#>  8:          lawyer  0.23770281      Male
#>  9:       scientist  0.16998545      Male
#> 10:          doctor  0.16863686      Male
#> 11:    psychologist  0.03023576      Male
#> 12:    investigator  0.01185508      Male
#> 13:      consultant -0.05517316    Female
#> 14:      programmer -0.14379956    Female
#> 15:         teacher -0.79991587    Female
#> 16:           clerk -1.03482346    Female
#> 17:       counselor -1.19585216    Female
#> 18:     salesperson -1.20660146    Female
#> 19:       therapist -1.24482768    Female
#> 20: psychotherapist -1.32262404    Female
#> 21:           nurse -1.49061460    Female
#>              T_word    eff_size closer_to
#>