Relative Norm Distance (RND) analysis.

Tabulate data for Relative Norm Distance (RND; also known as Relative Euclidean Distance) analysis. This is an alternative method to Single-Category WEAT.

Usage

test_RND(data, T1, A1, A2, use.pattern = FALSE, labels, rev = FALSE)

Arguments

data

A data.table with variables word and vec loaded by data_wordvec_load.

T1

Target words of a single category (a vector of words or a pattern of regular expression).

A1, A2

Attribute words (a vector of words or a pattern of regular expression). Both must be specified.

use.pattern

Defaults to FALSE (using a vector of words). If you use regular expression in T1, T2, A1, and A2, please specify this argument as TRUE.

labels

Labels for target and attribute concepts (a named list), such as (the default) list(T1="Target", A1="Attrib1", A2="Attrib2").

rev

Reverse the results? Defaults to FALSE, which means that a positive (vs. negative) value of RND indicates the target words are more associated with A2 than A1. If reversed (TRUE), then the interpretation of RND will also be reversed.

Value

A list of objects:

words.valid

valid (actually matched) words

data.rnd

data.table of (raw and relative) norm distances

code.diff

description for the difference between the two attribute concepts

eff.type

effect type: RND

eff.sum

sum of RND (a single value)

eff.interpretation

interpretation of the RND score

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

Garg, N., Schiebinger, L., Jurafsky, D., & Zou, J. (2018). Word embeddings quantify 100 years of gender and ethnic stereotypes. Proceedings of the National Academy of Sciences, 115(16), E3635--E3644.

Bhatia, N., & Bhatia, S. (2021). Changes in gender stereotypes over time: A computational analysis. Psychology of Women Quarterly, 45(1), 106--125.

See also

Examples

rnd = test_RND(
  demodata,
  T1=cc("
    architect, boss, leader, engineer, CEO, officer, manager,
    lawyer, scientist, doctor, psychologist, investigator,
    consultant, programmer, teacher, clerk, counselor,
    salesperson, therapist, psychotherapist, nurse"),
  A1=cc("male, man, boy, brother, he, him, his, son"),
  A2=cc("female, woman, girl, sister, she, her, hers, daughter"),
  labels=list(T1="Occupation", A1="Male", A2="Female"))
rnd
#> $words.valid
#> $words.valid$T1
#>  [1] "architect"       "boss"            "leader"          "engineer"       
#>  [5] "CEO"             "officer"         "manager"         "lawyer"         
#>  [9] "scientist"       "doctor"          "psychologist"    "investigator"   
#> [13] "consultant"      "programmer"      "teacher"         "clerk"          
#> [17] "counselor"       "salesperson"     "therapist"       "psychotherapist"
#> [21] "nurse"          
#> 
#> $words.valid$A1
#> [1] "male"    "man"     "boy"     "brother" "he"      "him"     "his"    
#> [8] "son"    
#> 
#> $words.valid$A2
#> [1] "female"   "woman"    "girl"     "sister"   "she"      "her"      "hers"    
#> [8] "daughter"
#> 
#> 
#> $data.rnd
#>              T_word norm_dist_A1 norm_dist_A2         rnd closer_to
#>  1:       architect    1.1375212    1.2425982 -0.10507699      Male
#>  2:            boss    1.0275216    1.1375335 -0.11001189      Male
#>  3:          leader    1.1046959    1.2062678 -0.10157189      Male
#>  4:        engineer    1.1232237    1.2163417 -0.09311807      Male
#>  5:             CEO    1.2349007    1.3002154 -0.06531466      Male
#>  6:         officer    1.0982037    1.1607926 -0.06258887      Male
#>  7:         manager    1.1710080    1.2215931 -0.05058515      Male
#>  8:          lawyer    1.0478646    1.1021396 -0.05427499      Male
#>  9:       scientist    1.1347233    1.1747094 -0.03998609      Male
#> 10:          doctor    0.9649817    1.0168649 -0.05188319      Male
#> 11:    psychologist    1.0798196    1.1154046 -0.03558499      Male
#> 12:    investigator    1.0950189    1.1297580 -0.03473911      Male
#> 13:      consultant    1.1722405    1.2016051 -0.02936453      Male
#> 14:      programmer    1.1723926    1.1993873 -0.02699467      Male
#> 15:         teacher    1.0275150    0.9987394  0.02877561    Female
#> 16:           clerk    1.0458337    1.0069546  0.03887911    Female
#> 17:       counselor    1.0890461    1.0645395  0.02450660    Female
#> 18:     salesperson    1.1233812    1.0829710  0.04041016    Female
#> 19:       therapist    1.0590592    1.0291238  0.02993539    Female
#> 20: psychotherapist    1.1153312    1.0656332  0.04969798    Female
#> 21:           nurse    1.0528372    0.9274327  0.12540452    Female
#>              T_word norm_dist_A1 norm_dist_A2         rnd closer_to
#> 
#> $code.diff
#> [1] "Occupation :: Male vs. Female"
#> 
#> $eff.type
#> [1] "Relative Norm Distance (RND)"
#> 
#> $eff.sum
#> [1] -0.5234857
#> 
#> $eff.interpretation
#> [1] "If RND < 0: Occupation is more associated with Male than Female"
#> [2] "If RND > 0: Occupation is more associated with Female than Male"
#>