Set up a Python environment for contextualized (dynamic) word embeddings, necessary for all text_* functions.

Install required Python packages in a conda environment (with defaults) and initialize the installed conda environment. Run once before your first use of any text_* function. No need to run this in the next time. Users may first need to manually install Anaconda or Miniconda.

Usage

text_env_setup(install = TRUE, initialize = TRUE, save = TRUE, ...)

Arguments

install

Run text::textrpp_install(). Defaults to TRUE.

initialize

Run text::textrpp_initialize(). Defaults to TRUE.

save

Save settings to .Rprofile for future use, so that you do not have to run this again after restarting R. Defaults to TRUE.

...

Other parameters passed to text::textrpp_install().

Details

The R package text (https://www.r-text.org/) enables users access to HuggingFace Transformers models in R, through the R package reticulate as an interface to Python and the Python packages torch and transformers. This function calls two functions in the text package.

Users who have problems in running this function should directly run the functions in text:

text::textrpp_install()

text::textrpp_install_virtualenv()

text::textrpp_uninstall()

text::textrpp_initialize()

See also

Examples

if (FALSE) {
text_env_setup()
}