Find the Top-N most similar words.

Find the Top-N most similar words, which replicates the results produced by the Python gensim module most_similar() function. (Exact replication of gensim requires the same word vectors data, not the demodata used here in examples.)

Usage

most_similar(data, x, topn = 10, keep = FALSE, above = NULL, verbose = TRUE)

Arguments

data

A data.table (of new class wordvec) loaded by data_wordvec_load.

x

Can be one of the following:

a single word: "China"

a list of words: c("king", "queen")

cc(" king , queen ; man | woman")

an R formula (~ xxx) specifying words that positively and negatively contribute to the similarity (for word analogy): ~ boy - he + she ~ king - man + woman ~ Beijing - China + Japan

topn

Top-N most similar words. Defaults to 10.

keep

Keep words specified in x in results? Defaults to FALSE.

above

Defaults to NULL. Can be one of the following:

a threshold value to find all words with cosine similarities higher than this value

a critical word to find all words with cosine similarities higher than that with this critical word

If both topn and above are specified, above wins.

verbose

Print information to the console? Defaults to TRUE.

Value

A data.table with the most similar words and their cosine similarities. The row number of each word in the raw data is also returned, which may help determine the relative word frequency in some cases.

Two attributes are appended to the returned data.table (see examples): wordvec and wordvec.formula. Users may extract them for further use.

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

See also

Examples

d = data_wordvec_normalize(demodata)
#> ✔ All word vectors have now been normalized.

most_similar(d, "China")
#> [Word Vector] =~ China
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:   Chinese 0.7678081    924
#>  2:   Beijing 0.7648461   2086
#>  3:    Taiwan 0.7081156   3011
#>  4:  Shanghai 0.6727434   3932
#>  5:  Shenzhen 0.6239033   7984
#>  6: Guangzhou 0.6223897   7986
#>  7:      yuan 0.6005429   4619
#>  8:     India 0.6004212    485
#>  9:     Japan 0.5967756    821
#> 10:        Li 0.5882897   7558
most_similar(d, c("king", "queen"))
#> [Word Vector] =~ king + queen
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:     royal 0.5985594   6754
#>  2:     Queen 0.5487501   4267
#>  3:      King 0.4662653   1449
#>  4:    legend 0.3730853   5428
#>  5:     crown 0.3673733   5174
#>  6: superstar 0.3652350   7218
#>  7:  champion 0.3587134   1509
#>  8:      lady 0.3581903   4631
#>  9:     lover 0.3425792   7947
#> 10:      pope 0.3420032   6388
most_similar(d, cc(" king , queen ; man | woman "))
#> [Word Vector] =~ king + queen + man + woman
#> (normalized to unit length)
#>            word   cos_sim row_id
#>  1:        girl 0.6387542   1257
#>  2:         boy 0.5966774   1379
#>  3:    teenager 0.5941694   4345
#>  4:        lady 0.5923386   4631
#>  5: grandmother 0.5081522   5364
#>  6:      mother 0.5000261    754
#>  7:       lover 0.4826535   7947
#>  8:      victim 0.4806608   1537
#>  9:   boyfriend 0.4717627   4811
#> 10:  girlfriend 0.4709186   3922

# the same as above:
most_similar(d, ~ China)
#> [Word Vector] =~ China
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:   Chinese 0.7678081    924
#>  2:   Beijing 0.7648461   2086
#>  3:    Taiwan 0.7081156   3011
#>  4:  Shanghai 0.6727434   3932
#>  5:  Shenzhen 0.6239033   7984
#>  6: Guangzhou 0.6223897   7986
#>  7:      yuan 0.6005429   4619
#>  8:     India 0.6004212    485
#>  9:     Japan 0.5967756    821
#> 10:        Li 0.5882897   7558
most_similar(d, ~ king + queen)
#> [Word Vector] =~ king + queen
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:     royal 0.5985594   6754
#>  2:     Queen 0.5487501   4267
#>  3:      King 0.4662653   1449
#>  4:    legend 0.3730853   5428
#>  5:     crown 0.3673733   5174
#>  6: superstar 0.3652350   7218
#>  7:  champion 0.3587134   1509
#>  8:      lady 0.3581903   4631
#>  9:     lover 0.3425792   7947
#> 10:      pope 0.3420032   6388
most_similar(d, ~ king + queen + man + woman)
#> [Word Vector] =~ king + queen + man + woman
#> (normalized to unit length)
#>            word   cos_sim row_id
#>  1:        girl 0.6387542   1257
#>  2:         boy 0.5966774   1379
#>  3:    teenager 0.5941694   4345
#>  4:        lady 0.5923386   4631
#>  5: grandmother 0.5081522   5364
#>  6:      mother 0.5000261    754
#>  7:       lover 0.4826535   7947
#>  8:      victim 0.4806608   1537
#>  9:   boyfriend 0.4717627   4811
#> 10:  girlfriend 0.4709186   3922

most_similar(d, ~ boy - he + she)
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>            word   cos_sim row_id
#>  1:        girl 0.8635271   1257
#>  2:       woman 0.6822032    558
#>  3:      mother 0.6530156    754
#>  4:    daughter 0.6431009   1078
#>  5:    teenager 0.6310561   4345
#>  6:       child 0.5933921    702
#>  7: grandmother 0.5800967   5364
#>  8:        teen 0.5755065   3542
#>  9:        baby 0.5672891   1838
#> 10:       girls 0.5599151   1195
most_similar(d, ~ Jack - he + she)
#> [Word Vector] =~ Jack - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#>  1:    Jane 0.6338590   5502
#>  2: Rebecca 0.6054167   6896
#>  3:   Julie 0.5988102   5721
#>  4:   Sarah 0.5867900   3717
#>  5:     Amy 0.5831094   4899
#>  6:   Susan 0.5812214   4141
#>  7:    Lisa 0.5766206   4521
#>  8:     Ann 0.5765654   4955
#>  9:   Alice 0.5649283   7340
#> 10:   Carol 0.5647291   5861
most_similar(d, ~ Rose - she + he)
#> [Word Vector] =~ Rose - she + he
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:  Leonard 0.4443190   6245
#>  2:   Martin 0.4206247   1425
#>  3:   Thomas 0.4078561   1214
#>  4:  Wallace 0.3833057   4350
#>  5:  Francis 0.3793813   5220
#>  6:  Johnson 0.3757961    761
#>  7:    Allen 0.3736620   2257
#>  8: Robinson 0.3725750   2698
#>  9:    Evans 0.3639734   3379
#> 10:   Duncan 0.3635448   4738

most_similar(d, ~ king - man + woman)
#> [Word Vector] =~ king - man + woman
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:    queen 0.7118192   7852
#>  2:    royal 0.4938203   6754
#>  3:    Queen 0.4346379   4267
#>  4:     King 0.3749903   1449
#>  5:      she 0.3341126     65
#>  6:     lady 0.3282869   4631
#>  7:   mother 0.3241257    754
#>  8:    crown 0.3164823   5174
#>  9:     hers 0.3073009   7981
#> 10: daughter 0.3021213   1078
most_similar(d, ~ Tokyo - Japan + China)
#> [Word Vector] =~ Tokyo - Japan + China
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:   Beijing 0.8216199   2086
#>  2:  Shanghai 0.7951419   3932
#>  3: Guangzhou 0.6529652   7986
#>  4:   Chinese 0.6439487    924
#>  5:  Shenzhen 0.6439113   7984
#>  6:     Seoul 0.5868835   5180
#>  7:      yuan 0.5821307   4619
#>  8:        Li 0.5712056   7558
#>  9:      Wang 0.5192575   7083
#> 10:    Moscow 0.5082187   3163
most_similar(d, ~ Beijing - China + Japan)
#> [Word Vector] =~ Beijing - China + Japan
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:     Tokyo 0.8115592   3017
#>  2:     Seoul 0.6568831   5180
#>  3:  Japanese 0.6475989   1562
#>  4: Pyongyang 0.5348969   6322
#>  5:   Bangkok 0.4677356   6510
#>  6:     Korea 0.4660699   3768
#>  7:       yen 0.4631333   2695
#>  8:    Taiwan 0.4330458   3011
#>  9:    Moscow 0.4217667   3163
#> 10: Guangzhou 0.4154183   7986

most_similar(d, "China", above=0.7)
#> [Word Vector] =~ China
#> (normalized to unit length)
#>       word   cos_sim row_id
#> 1: Chinese 0.7678081    924
#> 2: Beijing 0.7648461   2086
#> 3:  Taiwan 0.7081156   3011
most_similar(d, "China", above="Shanghai")
#> [Word Vector] =~ China
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:  Chinese 0.7678081    924
#> 2:  Beijing 0.7648461   2086
#> 3:   Taiwan 0.7081156   3011
#> 4: Shanghai 0.6727434   3932

# automatically normalized for more accurate results
ms = most_similar(demodata, ~ king - man + woman)
#> ! Results may be inaccurate if word vectors are not normalized.
#> ✔ All word vectors have now been normalized.
#> [Word Vector] =~ king - man + woman
#> (normalized to unit length)
ms
#>         word   cos_sim row_id
#>  1:    queen 0.7118192   7852
#>  2:    royal 0.4938203   6754
#>  3:    Queen 0.4346379   4267
#>  4:     King 0.3749903   1449
#>  5:      she 0.3341126     65
#>  6:     lady 0.3282869   4631
#>  7:   mother 0.3241257    754
#>  8:    crown 0.3164823   5174
#>  9:     hers 0.3073009   7981
#> 10: daughter 0.3021213   1078
str(ms)
#> Classes ‘wordvec’, ‘data.table’ and 'data.frame':	10 obs. of  3 variables:
#>  $ word   : chr  "queen" "royal" "Queen" "King" ...
#>  $ cos_sim: num  0.712 0.494 0.435 0.375 0.334 ...
#>  $ row_id : int  7852 6754 4267 1449 65 4631 754 5174 7981 1078
#>  - attr(*, ".internal.selfref")=<externalptr> 
#>  - attr(*, "dims")= int 300
#>  - attr(*, "normalized")= logi TRUE
#>  - attr(*, "wordvec")= num [1:300] -0.0055 -0.06705 -0.04519 0.03875 -0.00286 ...
#>  - attr(*, "wordvec.formula")=Class 'formula'  language ~king - man + woman
#>   .. ..- attr(*, ".Environment")=<environment: 0x55fb31efff88> 
attr(ms, "dims")
#> [1] 300
attr(ms, "normalized")
#> [1] TRUE
attr(ms, "wordvec.formula")
#> ~king - man + woman
#> <environment: 0x55fb31efff88>
attr(ms, "wordvec")
#>   [1] -0.0055029992 -0.0670502053 -0.0451917763  0.0387523404 -0.0028627068
#>   [6] -0.0311541918  0.0722744639 -0.0547962213 -0.0020958381  0.1267937027
#>  [11] -0.1177978229 -0.1116193754 -0.0432157999  0.0087648568 -0.0212594939
#>  [16] -0.0361886454 -0.0145757764 -0.0391255380  0.0236719225  0.0679695260
#>  [21]  0.0825006128  0.0179503178 -0.0102744498  0.0096805959  0.0097418516
#>  [26] -0.0583066547 -0.1043284916  0.0066451406  0.0774529645 -0.0630043398
#>  [31]  0.0176932143 -0.0357660002 -0.0127675640  0.0952755743 -0.1004202304
#>  [36] -0.0307725420  0.0191817489  0.0342936599 -0.0381161430 -0.0036010425
#>  [41] -0.0222736867 -0.0654523468  0.0581761410  0.0386649344  0.0625108822
#>  [46]  0.0144455335 -0.0298451219  0.0179820393 -0.0070950816  0.0410382461
#>  [51]  0.0193902936  0.0009054039 -0.0824461123  0.0896711960 -0.1071550823
#>  [56]  0.0090518215 -0.0452947722 -0.0242425101  0.0499748883  0.0017788248
#>  [61]  0.0896162255  0.0778694536  0.0043599153  0.0777055479  0.0090822422
#>  [66] -0.0272763870  0.0142941177  0.0595841030 -0.0188500428  0.0720441406
#>  [71]  0.0662750676  0.0333189751 -0.0275026411  0.0473774977  0.0302065682
#>  [76]  0.0363555415 -0.0305552844  0.0234168307  0.1017575023  0.0411732703
#>  [81]  0.0094623327  0.0278133126 -0.0067087575  0.0304040180 -0.0629132792
#>  [86]  0.0630248067 -0.0340541402  0.0388550926  0.0546983130  0.0218934638
#>  [91]  0.0433205797 -0.1001378189 -0.0860689768 -0.0450885649  0.0072867091
#>  [96]  0.0317569448  0.0645638429 -0.0160440819  0.0639956819 -0.0188932674
#> [101]  0.0053826412 -0.0191800667 -0.0261438897 -0.1247332173 -0.1010494892
#> [106]  0.0587875995 -0.1849463247 -0.0674423763  0.0657176846 -0.0205700926
#> [111]  0.0666514408  0.0815598320  0.0569012270 -0.0101953971  0.0484093475
#> [116] -0.1185521372  0.0030313936 -0.0933326263  0.0544090501  0.0927743854
#> [121]  0.0572259111  0.0394707134 -0.0271925329 -0.0265718034  0.0412028096
#> [126]  0.0438836936  0.0418592306  0.0034390046  0.0029262269  0.0134619940
#> [131]  0.0060545735  0.0513732397  0.0327279879  0.1054238844  0.0668562286
#> [136]  0.0769686804 -0.0227940919 -0.0422079731  0.0703340490 -0.0155774872
#> [141]  0.0372660027 -0.1356710009  0.0350093190  0.0236513594  0.0484942080
#> [146] -0.0400414708  0.0188626798  0.0274173715 -0.0486294492  0.0038126260
#> [151]  0.0133522101  0.0559058596  0.0022117502 -0.0837040985 -0.0849908637
#> [156] -0.0533659570 -0.0608346184  0.0319297479  0.0282798544  0.0078445029
#> [161] -0.0496067834 -0.0011413436  0.1585030627 -0.0257539005 -0.0556869160
#> [166] -0.0376501897  0.0119025195 -0.0722575073 -0.0134277815  0.0713891951
#> [171] -0.1004959954  0.0377733481  0.1079291741  0.0487607486  0.0176890786
#> [176] -0.0642323537  0.0939240962  0.0323515907  0.0062917470  0.0628913182
#> [181] -0.1168540632  0.0042852646 -0.1256198725 -0.0767676420  0.0159300785
#> [186]  0.0294246554  0.0809427655  0.0357663118 -0.0218166578  0.0638628914
#> [191]  0.0283086395 -0.0002039863 -0.0330815284  0.0067946163  0.0198031884
#> [196]  0.0572634491  0.0025753907 -0.0179017809  0.0275942861 -0.0879453025
#> [201] -0.0565598458  0.0206358787  0.0228473515 -0.0461282294 -0.0145729556
#> [206] -0.0143933569  0.0095678299  0.0631869552  0.0659021305  0.0042216349
#> [211]  0.0293649478  0.1033812547 -0.0516775577 -0.1253424596  0.0270994610
#> [216]  0.0385621533 -0.0036571112 -0.0278117809  0.0295294855  0.0024058853
#> [221] -0.0144733111  0.0180931805  0.0579378103 -0.0116669835  0.0087559466
#> [226] -0.0283575865 -0.0278965844 -0.0272467575  0.0342410954  0.0878225024
#> [231] -0.0741101358 -0.0684089597 -0.1740935244 -0.0044601583  0.0376029740
#> [236]  0.0221138776  0.0207988624 -0.0113099903 -0.0275133395 -0.0000199454
#> [241]  0.0563714279 -0.1243700668  0.0584321684  0.0294712221  0.0939427080
#> [246] -0.0011363452 -0.0734356129  0.0425710531  0.0715560074 -0.0782001568
#> [251]  0.0274794287  0.0326764105  0.0565690520 -0.0124150119  0.0545865281
#> [256]  0.0131866584 -0.1333997208 -0.0673208573 -0.0237475804 -0.0439831482
#> [261]  0.0048958304  0.0900008028 -0.0692655794  0.0177480897 -0.0185615675
#> [266] -0.0438003657  0.0037071559  0.0152207093  0.0955843067  0.1068395259
#> [271]  0.0044241142  0.1174178922 -0.0404572738  0.0672502162 -0.0424039690
#> [276] -0.1056443456 -0.0766069551  0.0035567801 -0.0748190562  0.0503630702
#> [281]  0.0909204843 -0.0668049051  0.0951348101 -0.0171847044  0.0625944510
#> [286] -0.0035357800 -0.0205028927 -0.0161002229 -0.0182009416  0.0601725192
#> [291] -0.0265177060 -0.0849906153 -0.0761131339 -0.1206830347  0.0004351351
#> [296]  0.0062921084 -0.0137075363  0.0623639418  0.0806824072  0.0592501645
# final word vector computed according to the formula