Train static word embeddings using the Word2Vec, GloVe, or FastText algorithm.

Train static word embeddings using the Word2Vec, GloVe, or FastText algorithm with multi-threading.

Usage

train_wordvec(
  text,
  method = c("word2vec", "glove", "fasttext"),
  dims = 300,
  window = 5,
  min.freq = 5,
  threads = 8,
  model = c("skip-gram", "cbow"),
  loss = c("ns", "hs"),
  negative = 5,
  subsample = 1e-04,
  learning = 0.05,
  ngrams = c(3, 6),
  x.max = 10,
  convergence = -1,
  stopwords = character(0),
  encoding = "UTF-8",
  tolower = FALSE,
  normalize = FALSE,
  iteration,
  tokenizer,
  remove,
  file.save,
  compress = "bzip2",
  verbose = TRUE
)

Arguments

text

A character vector of text, or a file path on disk containing text.

method

Training algorithm:

"word2vec" (default): using the word2vec package

"glove": using the rsparse and text2vec packages

"fasttext": using the fastTextR package

dims

Number of dimensions of word vectors to be trained. Common choices include 50, 100, 200, 300, and 500. Defaults to 300.

window

Window size (number of nearby words behind/ahead the current word). It defines how many surrounding words to be included in training: [window] words behind and [window] words ahead ([window]*2 in total). Defaults to 5.

min.freq

Minimum frequency of words to be included in training. Words that appear less than this value of times will be excluded from vocabulary. Defaults to 5 (take words that appear at least five times).

threads

Number of CPU threads used for training. A modest value produces the fastest training. Too many threads are not always helpful. Defaults to 8.

model

<Only for Word2Vec / FastText>

Learning model architecture:

"skip-gram" (default): Skip-Gram, which predicts surrounding words given the current word

"cbow": Continuous Bag-of-Words, which predicts the current word based on the context

loss

<Only for Word2Vec / FastText>

Loss function (computationally efficient approximation):

"ns" (default): Negative Sampling

"hs": Hierarchical Softmax

negative

<Only for Negative Sampling in Word2Vec / FastText>

Number of negative examples. Values in the range 5~20 are useful for small training datasets, while for large datasets the value can be as small as 2~5. Defaults to 5.

subsample

<Only for Word2Vec / FastText>

Subsampling of frequent words (threshold for occurrence of words). Those that appear with higher frequency in the training data will be randomly down-sampled. Defaults to 0.0001 (1e-04).

learning

<Only for Word2Vec / FastText>

Initial (starting) learning rate, also known as alpha. Defaults to 0.05.

ngrams

<Only for FastText>

Minimal and maximal ngram length. Defaults to c(3, 6).

x.max

<Only for GloVe>

Maximum number of co-occurrences to use in the weighting function. Defaults to 10.

convergence

<Only for GloVe>

Convergence tolerance for SGD iterations. Defaults to -1.

stopwords

<Only for Word2Vec / GloVe>

A character vector of stopwords to be excluded from training.

encoding

Text encoding. Defaults to "UTF-8".

tolower

Convert all upper-case characters to lower-case? Defaults to FALSE.

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See normalize.

iteration

Number of training iterations. More iterations makes a more precise model, but computational cost is linearly proportional to iterations. Defaults to 5 for Word2Vec and FastText while 10 for GloVe.

tokenizer

Function used to tokenize the text. Defaults to text2vec::word_tokenizer.

remove

Strings (in regular expression) to be removed from the text. Defaults to "_|'|<br/>|<br />|e\\.g\\.|i\\.e\\.". You may turn off this by specifying remove=NULL.

file.save

File name of to-be-saved R data (must be .RData).

compress

Compression method for the saved file. Defaults to "bzip2".

Options include:

1 or "gzip": modest file size (fastest)

2 or "bzip2": small file size (fast)

3 or "xz": minimized file size (slow)

verbose

Print information to the console? Defaults to TRUE.

Value

A wordvec (data.table) with three variables: word, vec, freq.

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

All-in-one package:

https://CRAN.R-project.org/package=wordsalad

Word2Vec:

https://code.google.com/archive/p/word2vec/

https://CRAN.R-project.org/package=word2vec

https://github.com/maxoodf/word2vec

GloVe:

https://nlp.stanford.edu/projects/glove/

https://text2vec.org/glove.html

https://CRAN.R-project.org/package=text2vec

https://CRAN.R-project.org/package=rsparse

FastText:

https://fasttext.cc/

https://CRAN.R-project.org/package=fastTextR

See also

Examples

review = text2vec::movie_review  # a data.frame'
text = review$review

## Note: All the examples train 50 dims for faster code check.

## Word2Vec (SGNS)
dt1 = train_wordvec(
  text,
  method="word2vec",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      Word2Vec (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14205 unique tokens (time cost = 11 secs)

dt1
#> # wordvec (data.table): [14205 × 3] (normalized)
#>                 word                     vec  freq
#>     1:           the [ 0.2487, ...<50 dims>] 58797
#>     2:           and [ 0.1583, ...<50 dims>] 32193
#>     3:             a [ 0.1515, ...<50 dims>] 31783
#>     4:            of [ 0.1416, ...<50 dims>] 29142
#>     5:            to [-0.0598, ...<50 dims>] 27218
#> ------                                            
#> 14201:        parrot [ 0.3251, ...<50 dims>]     5
#> 14202:          Lori [ 0.2214, ...<50 dims>]     5
#> 14203:      shambles [ 0.1304, ...<50 dims>]     5
#> 14204: comprehension [ 0.1816, ...<50 dims>]     5
#> 14205:        drunks [ 0.1345, ...<50 dims>]     5
most_similar(dt1, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:      ive 0.8576512    124
#>  2:    Youve 0.8239092   1542
#>  3:   lately 0.8032280   1564
#>  4: Possibly 0.7856805   2573
#>  5:   Having 0.7813926   3487
#>  6:    WORST 0.7799947   4275
#>  7:     weve 0.7791783   5137
#>  8:     EVER 0.7693427   5989
#>  9:     ever 0.7667641   6230
#> 10: funniest 0.7646575   6904
most_similar(dt1, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1: woman 0.8166134    260
#> 2:  lady 0.7283297    299
#> 3:  girl 0.7254774   1140
#> 4: witch 0.7023769   2513
#> 5: grows 0.6982635   3813
most_similar(dt1, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8299139    260
#> 2: woman 0.8147234    299
#> 3:  aged 0.7310625    674
#> 4:   kid 0.7290537   1140
#> 5:  lady 0.7135632   1839

## GloVe
dt2 = train_wordvec(
  text,
  method="glove",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      GloVe
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: N/A
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  10 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 14 secs)

dt2
#> # wordvec (data.table): [14207 × 3] (normalized)
#>            word                     vec  freq
#>     1:      the [-0.0192, ...<50 dims>] 58797
#>     2:      and [-0.0206, ...<50 dims>] 32193
#>     3:        a [-0.0959, ...<50 dims>] 31783
#>     4:       of [-0.0633, ...<50 dims>] 29142
#>     5:       to [-0.0395, ...<50 dims>] 27218
#> ------                                       
#> 14203:      yea [ 0.2047, ...<50 dims>]     5
#> 14204:   yearly [ 0.0100, ...<50 dims>]     5
#> 14205: yearning [ 0.0094, ...<50 dims>]     5
#> 14206:   yelled [ 0.3670, ...<50 dims>]     5
#> 14207:      yer [-0.0591, ...<50 dims>]     5
most_similar(dt2, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>        word   cos_sim row_id
#>  1:    seen 0.9458566     74
#>  2:    ever 0.8949125     91
#>  3:   heard 0.7572082    110
#>  4:   worst 0.7459619    124
#>  5:   since 0.7444330    261
#>  6:   youve 0.7180144    262
#>  7:  havent 0.7140256    305
#>  8: watched 0.6914957    515
#>  9:  movies 0.6905510    767
#> 10:    been 0.6826533    950
most_similar(dt2, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1: woman 0.8527515    198
#> 2: young 0.7462366    260
#> 3:  girl 0.7443405    299
#> 4:   hit 0.7399315    523
#> 5: child 0.7371599    594
most_similar(dt2, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8112230    153
#> 2: young 0.7551295    198
#> 3: woman 0.7144566    260
#> 4: named 0.7127865    299
#> 5:   man 0.6799258    867

## FastText
dt3 = train_wordvec(
  text,
  method="fasttext",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      FastText (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 24 secs)

dt3
#> # wordvec (data.table): [14207 × 3] (normalized)
#>                word                     vec  freq
#>     1:          the [-0.0034, ...<50 dims>] 58797
#>     2:          and [ 0.1839, ...<50 dims>] 32193
#>     3:            a [ 0.0591, ...<50 dims>] 31783
#>     4:           of [ 0.1131, ...<50 dims>] 29142
#>     5:           to [ 0.0644, ...<50 dims>] 27218
#> ------                                           
#> 14203:        spray [ 0.1085, ...<50 dims>]     5
#> 14204: disabilities [ 0.1620, ...<50 dims>]     5
#> 14205:        crook [ 0.0344, ...<50 dims>]     5
#> 14206:     Syndrome [ 0.0616, ...<50 dims>]     5
#> 14207:      snipers [ 0.0821, ...<50 dims>]     5
most_similar(dt3, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>           word   cos_sim row_id
#>  1:       seen 0.8406447    110
#>  2:       Weve 0.8239997    765
#>  3:      Youve 0.8131417    945
#>  4:         ve 0.7669521   3105
#>  5:      WORST 0.7665118   3494
#>  6:      youve 0.7627156   5898
#>  7:        ive 0.7502564   6913
#>  8: beforehand 0.7398510   7108
#>  9:    funnier 0.7361102   9171
#> 10:     havent 0.7335308  12894
most_similar(dt3, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:    woman 0.8652028    261
#> 2: salesman 0.7751966    299
#> 3:     girl 0.7533137   5594
#> 4:   madman 0.7501868   6553
#> 5: henchman 0.7478125  12263
most_similar(dt3, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:     girl 0.7762483    261
#> 2:    woman 0.7323415    299
#> 3:     boys 0.6908813    676
#> 4:      kid 0.6901126   1045
#> 5: teenager 0.6742148   2364