Train static word vectors using the Word2Vec, GloVe, or FastText algorithm.

Train static word vectors using the Word2Vec, GloVe, or FastText algorithm.

Usage

train_wordvec(
  x,
  tokenizer,
  remove,
  method = c("word2vec", "glove", "fasttext"),
  dims = 300,
  window = 5,
  min.count = 5,
  iteration = 5,
  threads = 8,
  model = c("skip-gram", "cbow"),
  loss = c("ns", "hs"),
  negative = 5,
  subsample,
  learning = 0.05,
  ngrams = c(3, 6),
  stopwords = character(),
  encoding = "UTF-8",
  normalize = FALSE,
  file.save = NULL
)

Arguments

x

A file path on disk containing the text or a character vector of the text.

tokenizer

Function used to tokenize the text. Defaults to text2vec::word_tokenizer.

remove

Strings (in regular expression) to be removed from the text. Defaults to "_|'\\w+|<br/>|<br />|e\\.g\\.|i\\.e\\.". You may turn off this by specifying remove=NULL.

method

Training algorithm:

"word2vec" (default): using the word2vec package

"glove": using the rsparse package

"fasttext": using the fastTextR package

dims

Number of dimensions of word vectors to be trained. Common choices include 50, 100, 200, 300, and 500. Defaults to 300.

window

Window size (number of nearby words before/after the current word). It defines how many surrounding words to be included in training: [window] words behind and [window] words ahead (thus [window]*2 in total). Defaults to 5.

min.count

Minimum frequency of words to be included in training. Words that appear less than this value of times will be excluded from vocabulary. Defaults to 5.

iteration

Number of training iterations. More iterations makes a more precise model, but computational cost is linearly proportional to iterations. Defaults to 5.

threads

Number of CPU threads used for training. A modest value produces the fastest training. Too many threads are not always helpful. Defaults to 8.

model

<Only for Word2Vec / FastText>

Learning model architecture:

"skip-gram" (default): Skip-Gram, which predicts surrounding words given the current word

"cbow": Continuous Bag-of-Words, which predicts the current word based on the context

loss

<Only for Word2Vec / FastText>

Loss function (computationally efficient approximation):

"ns" (default): Negative Sampling

"hs": Hierarchical Softmax

negative

<Only for Negative Sampling in Word2Vec / FastText>

Number of negative examples. Values in the range 5~20 are useful for small training datasets, while for large datasets the value can be as small as 2~5. Defaults to 5.

subsample

<Only for Word2Vec / FastText>

Subsampling of frequent words (threshold for occurrence of words). Those that appear with higher frequency in the training data will be randomly down-sampled. Defaults to 0.001 for Word2Vec and 0.0001 for FastText.

learning

<Only for Word2Vec / FastText>

Initial (starting) learning rate, also known as alpha. Defaults to 0.05.

ngrams

<Only for FastText>

Minimal and maximal ngram length. Defaults to c(3, 6).

stopwords

<Only for Word2Vec / GloVe>

A character vector of stopwords to be excluded from training.

encoding

Text encoding of x and stop_words. Defaults to "UTF-8".

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See data_wordvec_normalize.

file.save

File name of to-be-saved R data (must be .RData).

Value

A data.table with two variables:

word

words

vec

raw or normalized word vectors

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

All-in-one package:

https://CRAN.R-project.org/package=wordsalad

Word2Vec:

https://CRAN.R-project.org/package=word2vec

https://github.com/maxoodf/word2vec

GloVe:

https://CRAN.R-project.org/package=text2vec

https://CRAN.R-project.org/package=rsparse

FastText:

https://CRAN.R-project.org/package=fastTextR

See also

Examples

text = text2vec::movie_review
# View(text)

## Word2Vec (SGNS)
dt1 = train_wordvec(
  text$review,
  method="word2vec",
  model="skip-gram", loss="ns",
  dims=50, window=5)  # 50 dims for faster code check
#> √ Tokenized: 70101 sentences (time cost = 3 secs)
#> * Text corpus: 5207469 characters, 1183760 tokens (roughly words)
#> * Training model info:
#> - Method:      word2vec
#> - Dimensions:  50
#> - Window size: 5
#> - Subsampling: 0.001
#> - Min. freq.:  5
#> Training...
#> √ Word vectors trained: 14069 valid tokens (time cost = 18 secs)
most_similar(dt1, ~ man - he + she)
#> * Results may be inaccurate if word vectors are not normalized.
#> √ All word vectors have now been normalized.
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>            word   cos_sim row_id
#>  1:       woman 0.8939551    242
#>  2:        girl 0.8416112    286
#>  3:        Lena 0.7263495   6705
#>  4:         boy 0.7170201    454
#>  5: businessman 0.7126746   6126
#>  6:     wealthy 0.7003508   3144
#>  7:        brat 0.7001061   6837
#>  8:      Cassie 0.6993795   8284
#>  9:    Daniella 0.6992856  11873
#> 10:        deaf 0.6988062   5593
most_similar(dt1, ~ boy - he + she)
#> * Results may be inaccurate if word vectors are not normalized.
#> √ All word vectors have now been normalized.
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:     girl 0.8551921    286
#>  2:    woman 0.7829413    242
#>  3:   Cassie 0.6777951   8284
#>  4: pregnant 0.6763666   2622
#>  5: daughter 0.6738037    582
#>  6:    fever 0.6712357  11781
#>  7:    widow 0.6663015   5271
#>  8:     aged 0.6657004   1829
#>  9:    young 0.6650897    196
#> 10:   lonely 0.6627248   2323

## FastText
dt3 = train_wordvec(
  text$review,
  method="fasttext",
  model="skip-gram", loss="ns",
  dims=50, window=5)  # 50 dims for faster code check
#> √ Tokenized: 70101 sentences (time cost = 2 secs)
#> * Text corpus: 5207469 characters, 1183760 tokens (roughly words)
#> * Training model info:
#> - Method:      fasttext
#> - Dimensions:  50
#> - Window size: 5
#> - Subsampling: 1e-04
#> - Min. freq.:  5
#> Training...
#> √ Word vectors trained: 14071 valid tokens (time cost = 24 secs)
most_similar(dt3, ~ man - he + she)
#> * Results may be inaccurate if word vectors are not normalized.
#> √ All word vectors have now been normalized.
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:    woman 0.8343923    242
#>  2: salesman 0.7705178   5756
#>  3:  caveman 0.7330798  12520
#>  4:   madman 0.7146455   6701
#>  5:     lady 0.7145583   1051
#>  6: henchman 0.7118803  12113
#>  7:  Mexican 0.6918008   3504
#>  8:     loan 0.6858231   9437
#>  9: Batwoman 0.6851408  11320
#> 10:    Lyman 0.6769829  11457
most_similar(dt3, ~ boy - he + she)
#> * Results may be inaccurate if word vectors are not normalized.
#> √ All word vectors have now been normalized.
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:     girl 0.8271082    285
#>  2:    woman 0.7305905    242
#>  3:     lady 0.6960023   1051
#>  4:     boys 0.6959241   1162
#>  5:    girls 0.6843334    554
#>  6: youngest 0.6730038   5402
#>  7:    older 0.6685612    926
#>  8: children 0.6594460    431
#>  9:    child 0.6566291    497
#> 10:      kid 0.6518815    643