Train word vectors using the Word2Vec, GloVe, or FastText algorithm.

Train word vectors using the Word2Vec, GloVe, or FastText algorithm with multi-threading.

Usage

train_wordvec(
  text,
  method = c("word2vec", "glove", "fasttext"),
  dims = 300,
  window = 5,
  min.freq = 5,
  threads = 8,
  model = c("skip-gram", "cbow"),
  loss = c("ns", "hs"),
  negative = 5,
  subsample = 1e-04,
  learning = 0.05,
  ngrams = c(3, 6),
  x.max = 10,
  convergence = -1,
  stopwords = character(0),
  encoding = "UTF-8",
  tolower = FALSE,
  normalize = FALSE,
  iteration,
  tokenizer,
  remove,
  file.save,
  compress = "bzip2",
  verbose = TRUE
)

Arguments

text

A character vector of the text, or a file path on disk containing the text.

method

Training algorithm:

"word2vec" (default): using the word2vec package

"glove": using the rsparse and text2vec packages

"fasttext": using the fastTextR package

dims

Number of dimensions of word vectors to be trained. Common choices include 50, 100, 200, 300, and 500. Defaults to 300.

window

Window size (number of nearby words behind/ahead the current word). It defines how many surrounding words to be included in training: [window] words behind and [window] words ahead ([window]*2 in total). Defaults to 5.

min.freq

Minimum frequency of words to be included in training. Words that appear less than this value of times will be excluded from vocabulary. Defaults to 5 (take words that appear at least five times).

threads

Number of CPU threads used for training. A modest value produces the fastest training. Too many threads are not always helpful. Defaults to 8.

model

<Only for Word2Vec / FastText>

Learning model architecture:

"skip-gram" (default): Skip-Gram, which predicts surrounding words given the current word

"cbow": Continuous Bag-of-Words, which predicts the current word based on the context

loss

<Only for Word2Vec / FastText>

Loss function (computationally efficient approximation):

"ns" (default): Negative Sampling

"hs": Hierarchical Softmax

negative

<Only for Negative Sampling in Word2Vec / FastText>

Number of negative examples. Values in the range 5~20 are useful for small training datasets, while for large datasets the value can be as small as 2~5. Defaults to 5.

subsample

<Only for Word2Vec / FastText>

Subsampling of frequent words (threshold for occurrence of words). Those that appear with higher frequency in the training data will be randomly down-sampled. Defaults to 0.0001 (1e-04).

learning

<Only for Word2Vec / FastText>

Initial (starting) learning rate, also known as alpha. Defaults to 0.05.

ngrams

<Only for FastText>

Minimal and maximal ngram length. Defaults to c(3, 6).

x.max

<Only for GloVe>

Maximum number of co-occurrences to use in the weighting function. Defaults to 10.

convergence

<Only for GloVe>

Convergence tolerance for SGD iterations. Defaults to -1.

stopwords

<Only for Word2Vec / GloVe>

A character vector of stopwords to be excluded from training.

encoding

Text encoding. Defaults to "UTF-8".

tolower

Convert all upper-case characters to lower-case? Defaults to FALSE.

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See data_wordvec_normalize.

iteration

Number of training iterations. More iterations makes a more precise model, but computational cost is linearly proportional to iterations. Defaults to 5 for Word2Vec and FastText while 10 for GloVe.

tokenizer

Function used to tokenize the text. Defaults to text2vec::word_tokenizer.

remove

Strings (in regular expression) to be removed from the text. Defaults to "_|'|<br/>|<br />|e\\.g\\.|i\\.e\\.". You may turn off this by specifying remove=NULL.

file.save

File name of to-be-saved R data (must be .RData).

compress

Compression method for the saved file. Defaults to "bzip2".

Options include:

1 or "gzip": modest file size (fastest)

2 or "bzip2": small file size (fast)

3 or "xz": minimized file size (slow)

verbose

Print information to the console? Defaults to TRUE.

Value

A data.table (of new class wordvec) with two variables:

word

words (tokens)

vec

raw or normalized word vectors

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

All-in-one package:

https://CRAN.R-project.org/package=wordsalad

Word2Vec:

https://code.google.com/archive/p/word2vec/

https://CRAN.R-project.org/package=word2vec

https://github.com/maxoodf/word2vec

GloVe:

https://nlp.stanford.edu/projects/glove/

https://text2vec.org/glove.html

https://CRAN.R-project.org/package=text2vec

https://CRAN.R-project.org/package=rsparse

FastText:

https://fasttext.cc/

https://CRAN.R-project.org/package=fastTextR

See also

Examples

review = text2vec::movie_review  # a data.frame'
text = review$review

## Note: All the examples train 50 dims for faster code check.

## Word2Vec (SGNS)
dt1 = train_wordvec(
  text,
  method="word2vec",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      Word2Vec (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14205 unique tokens (time cost = 12 secs)
#> ✔ All word vectors have now been normalized.

as_tibble(dt1)  # just check
#> # A tibble: 14,205 × 3
#>    word  vec         freq
#>    <chr> <list>     <int>
#>  1 the   <dbl [50]> 58797
#>  2 and   <dbl [50]> 32193
#>  3 a     <dbl [50]> 31783
#>  4 of    <dbl [50]> 29142
#>  5 to    <dbl [50]> 27218
#>  6 is    <dbl [50]> 21447
#>  7 in    <dbl [50]> 17861
#>  8 I     <dbl [50]> 14703
#>  9 that  <dbl [50]> 13618
#> 10 it    <dbl [50]> 13453
#> # … with 14,195 more rows
most_similar(dt1, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>           word   cos_sim row_id
#>  1:        ive 0.8887614   3487
#>  2:       seen 0.8210021    110
#>  3:  Bollywood 0.8104229   3585
#>  4:       ever 0.8068716    124
#>  5:     lately 0.8057448   5137
#>  6:       Weve 0.8000092   6928
#>  7:      Hands 0.7974401   7224
#>  8: criticized 0.7969465   8457
#>  9:      WORST 0.7914292   6904
#> 10:      Youve 0.7896730   5989
most_similar(dt1, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>      word   cos_sim row_id
#> 1:  woman 0.8553832    260
#> 2:   girl 0.7527926    299
#> 3:   lady 0.7464193   1140
#> 4: lonely 0.7252392   2338
#> 5:  young 0.7112162    197
most_similar(dt1, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:   kid 0.7647311    674
#> 2:  girl 0.7611924    299
#> 3: woman 0.7435880    260
#> 4: witch 0.7084293   2513
#> 5:  aged 0.7014026   1839

## GloVe
dt2 = train_wordvec(
  text,
  method="glove",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      GloVe
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: N/A
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  10 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 16 secs)
#> ✔ All word vectors have now been normalized.

as_tibble(dt2)  # just check
#> # A tibble: 14,207 × 3
#>    word  vec         freq
#>    <chr> <list>     <int>
#>  1 the   <dbl [50]> 58797
#>  2 and   <dbl [50]> 32193
#>  3 a     <dbl [50]> 31783
#>  4 of    <dbl [50]> 29142
#>  5 to    <dbl [50]> 27218
#>  6 is    <dbl [50]> 21447
#>  7 in    <dbl [50]> 17861
#>  8 I     <dbl [50]> 14703
#>  9 that  <dbl [50]> 13618
#> 10 it    <dbl [50]> 13453
#> # … with 14,197 more rows
most_similar(dt2, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>        word   cos_sim row_id
#>  1:    seen 0.9411882    110
#>  2:    ever 0.8944266    124
#>  3:   worst 0.7543924    261
#>  4:   heard 0.7518399    515
#>  5:   since 0.7300758    262
#>  6:   youve 0.7200879    950
#>  7: watched 0.7051912    305
#>  8:  movies 0.6895728     91
#>  9: already 0.6770569    468
#> 10:    been 0.6765670     74
most_similar(dt2, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1: woman 0.8718137    260
#> 2: young 0.7610926    198
#> 3:  girl 0.7610592    299
#> 4:   hit 0.7481330    594
#> 5: child 0.7409089    523
most_similar(dt2, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8163652    299
#> 2: young 0.7775646    198
#> 3: woman 0.7455849    260
#> 4: named 0.7199464    867
#> 5:   man 0.7007369    153

## FastText
dt3 = train_wordvec(
  text,
  method="fasttext",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      FastText (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 26 secs)
#> ✔ All word vectors have now been normalized.

as_tibble(dt3)  # just check
#> # A tibble: 14,207 × 3
#>    word  vec         freq
#>    <chr> <list>     <int>
#>  1 the   <dbl [50]> 58797
#>  2 and   <dbl [50]> 32193
#>  3 a     <dbl [50]> 31783
#>  4 of    <dbl [50]> 29142
#>  5 to    <dbl [50]> 27218
#>  6 is    <dbl [50]> 21447
#>  7 in    <dbl [50]> 17861
#>  8 I     <dbl [50]> 14703
#>  9 that  <dbl [50]> 13618
#> 10 it    <dbl [50]> 13453
#> # … with 14,197 more rows
most_similar(dt3, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>         word   cos_sim row_id
#>  1:    Youve 0.8343096   5898
#>  2:     Weve 0.8205638   6913
#>  3:     seen 0.8151530    110
#>  4:    youve 0.7832114    945
#>  5:       ve 0.7762302  12894
#>  6:    WORST 0.7755833   7108
#>  7:  funnier 0.7494815   3105
#>  8:  Columbo 0.7456561   3250
#>  9: funniest 0.7444905   1539
#> 10:      ive 0.7390809   3494
most_similar(dt3, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:    woman 0.8709031    261
#> 2:     girl 0.7755713    299
#> 3: salesman 0.7740031   5594
#> 4: henchman 0.7690305  12263
#> 5:  caveman 0.7432090  12884
most_similar(dt3, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:     girl 0.7899726    299
#> 2:    woman 0.7634606    261
#> 3:     boys 0.7044068   1045
#> 4: teenager 0.6781507   2364
#> 5:      kid 0.6713795    676