Train static word vectors using the Word2Vec, GloVe, or FastText algorithm.

Train static word vectors using the Word2Vec, GloVe, or FastText algorithm.

Usage

train_wordvec(
  x,
  tokenizer = text2vec::space_tokenizer,
  method = c("word2vec", "glove", "fasttext"),
  dims = 300,
  window = 5,
  min.count = 5,
  iteration = 5,
  threads = 8,
  model = c("skip-gram", "cbow"),
  loss = c("ns", "hs"),
  negative = 5,
  subsample = 0.001,
  alpha = 0.05,
  split = utf8_split_default(),
  stopwords = character(),
  encoding = "UTF-8",
  normalize = FALSE,
  file.save = NULL
)

Arguments

x

A character vector of text or a file path on disk containing the text.

tokenizer

Function used to tokenize the text. Defaults to text2vec::space_tokenizer().

method

Training algorithm:

"word2vec" (default): using the word2vec package

"glove": using the rsparse package

"fasttext": using the fastTextR package

dims

Number of dimensions of word vectors to be trained. Common choices include 50, 100, 200, 300, and 500. Defaults to 300.

window

Window size (number of nearby words before/after the current word). It defines how many surrounding words to be included in training: [window] words behind and [window] words ahead (thus [window]*2 in total). Defaults to 5.

min.count

Minimum frequency of words to be included in training. Words that appear less than this value of times will be excluded from vocabulary. Defaults to 5.

iteration

Number of training iterations. More iterations makes a more precise model, but computational cost is linearly proportional to iterations. Defaults to 5.

threads

Number of CPU threads used for training. A modest value produces the fastest training. Too many threads are not always helpful. Defaults to 8.

model

<Only for Word2Vec / FastText>

Learning model architecture:

"skip-gram" (default): Skip-Gram, which predicts surrounding words given the current word

"cbow": Continuous Bag-of-Words, which predicts the current word based on the context

loss

<Only for Word2Vec / FastText>

Loss function (computationally efficient approximation):

"ns" (default): Negative Sampling

"hs": Hierarchical Softmax

negative

<Only for Negative Sampling in Word2Vec / FastText>

Number of negative examples. Values in the range 5~20 are useful for small training datasets, while for large datasets the value can be as small as 2~5. Defaults to 5.

subsample

<Only for Word2Vec / FastText>

Subsampling of frequent words (threshold for occurrence of words). Those that appear with higher frequency in the training data will be randomly down-sampled. Defaults to 0.001.

alpha

<Only for Word2Vec / FastText>

Initial (starting) learning rate, also known as alpha. Defaults to 0.05.

split

<Only for Word2Vec>

A character vector of length 2: (1) the first element indicates how to split words and (2) the second element indicates how to split sentences. Defaults to a pre-defined utf8_split_default.

stopwords

<Only for Word2Vec / GloVe>

A character vector of stopwords to be excluded from training.

encoding

Text encoding of x and stop_words. Defaults to "UTF-8".

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See data_wordvec_normalize.

file.save

File name of to-be-saved R data (must be .RData).

Value

A data.table with two variables:

word

words

vec

raw or normalized word vectors

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

All-in-one package:

https://CRAN.R-project.org/package=wordsalad

Word2Vec:

https://CRAN.R-project.org/package=word2vec

https://github.com/maxoodf/word2vec

GloVe:

https://CRAN.R-project.org/package=text2vec

https://CRAN.R-project.org/package=rsparse

FastText:

https://CRAN.R-project.org/package=fastTextR

Examples

text = text2vec::movie_review
# View(text)

## Word2Vec (SGNS)
dt1 = train_wordvec(
  text$review,
  method="word2vec",
  model="skip-gram", loss="ns",
  dims=50, window=5)  # 50 dims for faster code check
#> * Text corpus: 5000 rows, 1205783 tokens (roughly words)
#> * Training model info:
#> - Method:      word2vec
#> - Dimensions:  50
#> - Window size: 5
#> - Subsampling: 0.001
#> - Min. freq.:  5
#> √ Word vectors have been trained (time cost = 19 secs)
most_similar(dt1, ~ father - he + she)
#> * Results may be inaccurate if word vectors are not normalized.
#> √ All word vectors have now been normalized.
#> [Word Vector] =~ father - he + she
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:    mother 0.8506691    502
#>  2:   husband 0.8441984    649
#>  3:    sister 0.8106275    811
#>  4:  daughter 0.8090412    603
#>  5:       son 0.7724012    480
#>  6:       mom 0.7510509   1864
#>  7:      aunt 0.7430656   3480
#>  8:      wife 0.7429989    314
#>  9:    Claire 0.7338367   3498
#> 10: boyfriend 0.7317490   1643