Train static word embeddings using the Word2Vec, GloVe, or FastText algorithm.

Train static word embeddings using the Word2Vec, GloVe, or FastText algorithm with multi-threading.

Usage

train_wordvec(
  text,
  method = c("word2vec", "glove", "fasttext"),
  dims = 300,
  window = 5,
  min.freq = 5,
  threads = 8,
  model = c("skip-gram", "cbow"),
  loss = c("ns", "hs"),
  negative = 5,
  subsample = 1e-04,
  learning = 0.05,
  ngrams = c(3, 6),
  x.max = 10,
  convergence = -1,
  stopwords = character(0),
  encoding = "UTF-8",
  tolower = FALSE,
  normalize = FALSE,
  iteration,
  tokenizer,
  remove,
  file.save,
  compress = "bzip2",
  verbose = TRUE
)

Arguments

text

A character vector of text, or a file path on disk containing text.

method

Training algorithm:

"word2vec" (default): using the word2vec package

"glove": using the rsparse and text2vec packages

"fasttext": using the fastTextR package

dims

Number of dimensions of word vectors to be trained. Common choices include 50, 100, 200, 300, and 500. Defaults to 300.

window

Window size (number of nearby words behind/ahead the current word). It defines how many surrounding words to be included in training: [window] words behind and [window] words ahead ([window]*2 in total). Defaults to 5.

min.freq

Minimum frequency of words to be included in training. Words that appear less than this value of times will be excluded from vocabulary. Defaults to 5 (take words that appear at least five times).

threads

Number of CPU threads used for training. A modest value produces the fastest training. Too many threads are not always helpful. Defaults to 8.

model

<Only for Word2Vec / FastText>

Learning model architecture:

"skip-gram" (default): Skip-Gram, which predicts surrounding words given the current word

"cbow": Continuous Bag-of-Words, which predicts the current word based on the context

loss

<Only for Word2Vec / FastText>

Loss function (computationally efficient approximation):

"ns" (default): Negative Sampling

"hs": Hierarchical Softmax

negative

<Only for Negative Sampling in Word2Vec / FastText>

Number of negative examples. Values in the range 5~20 are useful for small training datasets, while for large datasets the value can be as small as 2~5. Defaults to 5.

subsample

<Only for Word2Vec / FastText>

Subsampling of frequent words (threshold for occurrence of words). Those that appear with higher frequency in the training data will be randomly down-sampled. Defaults to 0.0001 (1e-04).

learning

<Only for Word2Vec / FastText>

Initial (starting) learning rate, also known as alpha. Defaults to 0.05.

ngrams

<Only for FastText>

Minimal and maximal ngram length. Defaults to c(3, 6).

x.max

<Only for GloVe>

Maximum number of co-occurrences to use in the weighting function. Defaults to 10.

convergence

<Only for GloVe>

Convergence tolerance for SGD iterations. Defaults to -1.

stopwords

<Only for Word2Vec / GloVe>

A character vector of stopwords to be excluded from training.

encoding

Text encoding. Defaults to "UTF-8".

tolower

Convert all upper-case characters to lower-case? Defaults to FALSE.

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See normalize.

iteration

Number of training iterations. More iterations makes a more precise model, but computational cost is linearly proportional to iterations. Defaults to 5 for Word2Vec and FastText while 10 for GloVe.

tokenizer

Function used to tokenize the text. Defaults to text2vec::word_tokenizer.

remove

Strings (in regular expression) to be removed from the text. Defaults to "_|'|<br/>|<br />|e\\.g\\.|i\\.e\\.". You may turn off this by specifying remove=NULL.

file.save

File name of to-be-saved R data (must be .RData).

compress

Compression method for the saved file. Defaults to "bzip2".

Options include:

1 or "gzip": modest file size (fastest)

2 or "bzip2": small file size (fast)

3 or "xz": minimized file size (slow)

verbose

Print information to the console? Defaults to TRUE.

Value

A wordvec (data.table) with three variables: word, vec, freq.

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

All-in-one package:

https://CRAN.R-project.org/package=wordsalad

Word2Vec:

https://code.google.com/archive/p/word2vec/

https://CRAN.R-project.org/package=word2vec

https://github.com/maxoodf/word2vec

GloVe:

https://nlp.stanford.edu/projects/glove/

https://text2vec.org/glove.html

https://CRAN.R-project.org/package=text2vec

https://CRAN.R-project.org/package=rsparse

FastText:

https://fasttext.cc/

https://CRAN.R-project.org/package=fastTextR

See also

Examples

review = text2vec::movie_review  # a data.frame'
text = review$review

## Note: All the examples train 50 dims for faster code check.

## Word2Vec (SGNS)
dt1 = train_wordvec(
  text,
  method="word2vec",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 2 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      Word2Vec (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14205 unique tokens (time cost = 11 secs)

dt1
#> # wordvec (data.table): [14205 × 3] (normalized)
#>                 word                     vec  freq
#>     1:           the [ 0.1920, ...<50 dims>] 58797
#>     2:           and [ 0.2892, ...<50 dims>] 32193
#>     3:             a [ 0.0581, ...<50 dims>] 31783
#>     4:            of [ 0.2895, ...<50 dims>] 29142
#>     5:            to [ 0.1533, ...<50 dims>] 27218
#> ------                                            
#> 14201:        parrot [ 0.1482, ...<50 dims>]     5
#> 14202:          Lori [ 0.0605, ...<50 dims>]     5
#> 14203:      shambles [ 0.0624, ...<50 dims>]     5
#> 14204: comprehension [ 0.1901, ...<50 dims>]     5
#> 14205:        drunks [ 0.1310, ...<50 dims>]     5
most_similar(dt1, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>          word   cos_sim row_id
#>  1:       ive 0.8977448   1542
#>  2:  Possibly 0.8302731   2644
#>  3:     WORST 0.8278717   3125
#>  4:      Weve 0.8126914   3487
#>  5:    recall 0.8087366   3585
#>  6:  funniest 0.8075806   5989
#>  7:     quote 0.7970962   6230
#>  8: Bollywood 0.7955555   6904
#>  9:     Youve 0.7932447   6928
#> 10:     Hands 0.7928162   7224
most_similar(dt1, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8568552    197
#> 2: woman 0.8037703    260
#> 3: child 0.7746318    299
#> 4: young 0.7736850    523
#> 5:  aged 0.7575521   1839
most_similar(dt1, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8238748    197
#> 2:  aged 0.7680243    299
#> 3:   kid 0.7430376    523
#> 4: young 0.7241408    674
#> 5: child 0.7240791   1839

## GloVe
dt2 = train_wordvec(
  text,
  method="glove",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 2 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      GloVe
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: N/A
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  10 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 12 secs)

dt2
#> # wordvec (data.table): [14207 × 3] (normalized)
#>            word                     vec  freq
#>     1:      the [-0.0168, ...<50 dims>] 58797
#>     2:      and [-0.0398, ...<50 dims>] 32193
#>     3:        a [-0.0984, ...<50 dims>] 31783
#>     4:       of [-0.0615, ...<50 dims>] 29142
#>     5:       to [-0.0241, ...<50 dims>] 27218
#> ------                                       
#> 14203:      yea [ 0.2069, ...<50 dims>]     5
#> 14204:   yearly [ 0.0145, ...<50 dims>]     5
#> 14205: yearning [-0.0030, ...<50 dims>]     5
#> 14206:   yelled [ 0.3390, ...<50 dims>]     5
#> 14207:      yer [-0.0524, ...<50 dims>]     5
most_similar(dt2, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>        word   cos_sim row_id
#>  1:    seen 0.9483907     91
#>  2:    ever 0.8908177    110
#>  3:   heard 0.7668436    124
#>  4:   worst 0.7549940    261
#>  5:   since 0.7329158    262
#>  6:  havent 0.7149096    305
#>  7: watched 0.7141680    468
#>  8:   youve 0.7075121    515
#>  9: already 0.6999404    767
#> 10:  movies 0.6854584    950
most_similar(dt2, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1: woman 0.8597315    198
#> 2: young 0.7540652    260
#> 3:  girl 0.7462352    299
#> 4:   hit 0.7344277    523
#> 5: child 0.7273571    594
most_similar(dt2, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8234241    153
#> 2: young 0.7721083    198
#> 3: woman 0.7378901    260
#> 4: named 0.7200721    299
#> 5:   man 0.7019027    867

## FastText
dt3 = train_wordvec(
  text,
  method="fasttext",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 2 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      FastText (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 25 secs)

dt3
#> # wordvec (data.table): [14207 × 3] (normalized)
#>                word                     vec  freq
#>     1:          the [ 0.0295, ...<50 dims>] 58797
#>     2:          and [ 0.1171, ...<50 dims>] 32193
#>     3:            a [ 0.0794, ...<50 dims>] 31783
#>     4:           of [ 0.0902, ...<50 dims>] 29142
#>     5:           to [ 0.0710, ...<50 dims>] 27218
#> ------                                           
#> 14203:        spray [ 0.1312, ...<50 dims>]     5
#> 14204: disabilities [ 0.1485, ...<50 dims>]     5
#> 14205:        crook [ 0.0663, ...<50 dims>]     5
#> 14206:     Syndrome [ 0.0502, ...<50 dims>]     5
#> 14207:      snipers [ 0.1097, ...<50 dims>]     5
most_similar(dt3, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>           word   cos_sim row_id
#>  1:       seen 0.8436856    110
#>  2:      Youve 0.8334989    945
#>  3:       Weve 0.8130754   3105
#>  4:         ve 0.7803378   3250
#>  5:      youve 0.7766163   3494
#>  6:      WORST 0.7738033   5898
#>  7:    Columbo 0.7529365   6913
#>  8:        ive 0.7511789   7108
#>  9:    funnier 0.7500011   9171
#> 10: beforehand 0.7387751  12894
most_similar(dt3, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:    woman 0.8810232    261
#> 2:     girl 0.7686986    299
#> 3:   madman 0.7625394   5594
#> 4: henchman 0.7576569   6553
#> 5: salesman 0.7550006  12263
most_similar(dt3, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:     girl 0.7757623    261
#> 2:    woman 0.7179198    299
#> 3:     boys 0.7074330    676
#> 4:      kid 0.7064313   1045
#> 5: teenager 0.6856899   2364