Train static word embeddings using the Word2Vec, GloVe, or FastText algorithm.

Train static word embeddings using the Word2Vec, GloVe, or FastText algorithm with multi-threading.

Usage

train_wordvec(
  text,
  method = c("word2vec", "glove", "fasttext"),
  dims = 300,
  window = 5,
  min.freq = 5,
  threads = 8,
  model = c("skip-gram", "cbow"),
  loss = c("ns", "hs"),
  negative = 5,
  subsample = 1e-04,
  learning = 0.05,
  ngrams = c(3, 6),
  x.max = 10,
  convergence = -1,
  stopwords = character(0),
  encoding = "UTF-8",
  tolower = FALSE,
  normalize = FALSE,
  iteration,
  tokenizer,
  remove,
  file.save,
  compress = "bzip2",
  verbose = TRUE
)

Arguments

text

A character vector of text, or a file path on disk containing text.

method

Training algorithm:

"word2vec" (default): using the word2vec package

"glove": using the rsparse and text2vec packages

"fasttext": using the fastTextR package

dims

Number of dimensions of word vectors to be trained. Common choices include 50, 100, 200, 300, and 500. Defaults to 300.

window

Window size (number of nearby words behind/ahead the current word). It defines how many surrounding words to be included in training: [window] words behind and [window] words ahead ([window]*2 in total). Defaults to 5.

min.freq

Minimum frequency of words to be included in training. Words that appear less than this value of times will be excluded from vocabulary. Defaults to 5 (take words that appear at least five times).

threads

Number of CPU threads used for training. A modest value produces the fastest training. Too many threads are not always helpful. Defaults to 8.

model

<Only for Word2Vec / FastText>

Learning model architecture:

"skip-gram" (default): Skip-Gram, which predicts surrounding words given the current word

"cbow": Continuous Bag-of-Words, which predicts the current word based on the context

loss

<Only for Word2Vec / FastText>

Loss function (computationally efficient approximation):

"ns" (default): Negative Sampling

"hs": Hierarchical Softmax

negative

<Only for Negative Sampling in Word2Vec / FastText>

Number of negative examples. Values in the range 5~20 are useful for small training datasets, while for large datasets the value can be as small as 2~5. Defaults to 5.

subsample

<Only for Word2Vec / FastText>

Subsampling of frequent words (threshold for occurrence of words). Those that appear with higher frequency in the training data will be randomly down-sampled. Defaults to 0.0001 (1e-04).

learning

<Only for Word2Vec / FastText>

Initial (starting) learning rate, also known as alpha. Defaults to 0.05.

ngrams

<Only for FastText>

Minimal and maximal ngram length. Defaults to c(3, 6).

x.max

<Only for GloVe>

Maximum number of co-occurrences to use in the weighting function. Defaults to 10.

convergence

<Only for GloVe>

Convergence tolerance for SGD iterations. Defaults to -1.

stopwords

<Only for Word2Vec / GloVe>

A character vector of stopwords to be excluded from training.

encoding

Text encoding. Defaults to "UTF-8".

tolower

Convert all upper-case characters to lower-case? Defaults to FALSE.

normalize

Normalize all word vectors to unit length? Defaults to FALSE. See normalize.

iteration

Number of training iterations. More iterations makes a more precise model, but computational cost is linearly proportional to iterations. Defaults to 5 for Word2Vec and FastText while 10 for GloVe.

tokenizer

Function used to tokenize the text. Defaults to text2vec::word_tokenizer.

remove

Strings (in regular expression) to be removed from the text. Defaults to "_|'|<br/>|<br />|e\\.g\\.|i\\.e\\.". You may turn off this by specifying remove=NULL.

file.save

File name of to-be-saved R data (must be .RData).

compress

Compression method for the saved file. Defaults to "bzip2".

Options include:

1 or "gzip": modest file size (fastest)

2 or "bzip2": small file size (fast)

3 or "xz": minimized file size (slow)

verbose

Print information to the console? Defaults to TRUE.

Value

A wordvec (data.table) with three variables: word, vec, freq.

Download

Download pre-trained word vectors data (.RData): https://psychbruce.github.io/WordVector_RData.pdf

References

All-in-one package:

https://CRAN.R-project.org/package=wordsalad

Word2Vec:

https://code.google.com/archive/p/word2vec/

https://CRAN.R-project.org/package=word2vec

https://github.com/maxoodf/word2vec

GloVe:

https://nlp.stanford.edu/projects/glove/

https://text2vec.org/glove.html

https://CRAN.R-project.org/package=text2vec

https://CRAN.R-project.org/package=rsparse

FastText:

https://fasttext.cc/

https://CRAN.R-project.org/package=fastTextR

See also

Examples

review = text2vec::movie_review  # a data.frame'
text = review$review

## Note: All the examples train 50 dims for faster code check.

## Word2Vec (SGNS)
dt1 = train_wordvec(
  text,
  method="word2vec",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 4 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      Word2Vec (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14205 unique tokens (time cost = 12 secs)

dt1
#> # wordvec (data.table): [14205 × 3] (normalized)
#>                 word                     vec  freq
#>     1:           the [ 0.0097, ...<50 dims>] 58797
#>     2:           and [-0.0260, ...<50 dims>] 32193
#>     3:             a [-0.1658, ...<50 dims>] 31783
#>     4:            of [-0.0163, ...<50 dims>] 29142
#>     5:            to [ 0.0113, ...<50 dims>] 27218
#> ------                                            
#> 14201:        parrot [ 0.1183, ...<50 dims>]     5
#> 14202:          Lori [ 0.0998, ...<50 dims>]     5
#> 14203:      shambles [ 0.0064, ...<50 dims>]     5
#> 14204: comprehension [ 0.0283, ...<50 dims>]     5
#> 14205:        drunks [ 0.0578, ...<50 dims>]     5
most_similar(dt1, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>           word   cos_sim row_id
#>  1:        ive 0.8723342    110
#>  2:     lately 0.8303237    124
#>  3:   funniest 0.8233828   1542
#>  4:       ever 0.8070427   2563
#>  5: criticized 0.8062371   3125
#>  6:       seen 0.7876970   3487
#>  7:  Bollywood 0.7844217   3585
#>  8:  favorites 0.7820451   4778
#>  9:      quote 0.7800325   5137
#> 10:     dozens 0.7717610   8457
most_similar(dt1, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>      word   cos_sim row_id
#> 1:  woman 0.7961274    260
#> 2:    boy 0.7464479    299
#> 3:   girl 0.7438613    478
#> 4:   lady 0.7361061   1140
#> 5: lonely 0.7034124   2338
most_similar(dt1, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.7439605    260
#> 2:  lady 0.7382876    299
#> 3: woman 0.7377241    674
#> 4:  aged 0.6676347   1140
#> 5:   kid 0.6616207   1839

## GloVe
dt2 = train_wordvec(
  text,
  method="glove",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 4 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      GloVe
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: N/A
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  10 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 19 secs)

dt2
#> # wordvec (data.table): [14207 × 3] (normalized)
#>            word                     vec  freq
#>     1:      the [ 0.0330, ...<50 dims>] 58797
#>     2:      and [ 0.0416, ...<50 dims>] 32193
#>     3:        a [-0.0542, ...<50 dims>] 31783
#>     4:       of [ 0.0044, ...<50 dims>] 29142
#>     5:       to [-0.0339, ...<50 dims>] 27218
#> ------                                       
#> 14203:      yea [ 0.1911, ...<50 dims>]     5
#> 14204:   yearly [-0.0276, ...<50 dims>]     5
#> 14205: yearning [-0.0080, ...<50 dims>]     5
#> 14206:   yelled [ 0.2891, ...<50 dims>]     5
#> 14207:      yer [-0.0759, ...<50 dims>]     5
most_similar(dt2, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>        word   cos_sim row_id
#>  1:    seen 0.9410974     91
#>  2:    ever 0.8905264    110
#>  3:   heard 0.7557552    124
#>  4:   worst 0.7493428    261
#>  5:   since 0.7441188    262
#>  6: watched 0.7027292    305
#>  7:   youve 0.6984664    468
#>  8:  movies 0.6891568    515
#>  9: already 0.6812694    767
#> 10:  havent 0.6807948    950
most_similar(dt2, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1: woman 0.8543943     34
#> 2:  girl 0.7602050    198
#> 3: young 0.7602017    260
#> 4:   hit 0.7476384    299
#> 5:   who 0.7212388    594
most_similar(dt2, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>     word   cos_sim row_id
#> 1:  girl 0.8059767    153
#> 2: young 0.7519598    198
#> 3: woman 0.7078108    260
#> 4: named 0.7071330    299
#> 5:   man 0.6959355    867

## FastText
dt3 = train_wordvec(
  text,
  method="fasttext",
  model="skip-gram",
  dims=50, window=5,
  normalize=TRUE)
#> ✔ Tokenized: 70105 sentences (time cost = 3 secs)
#> ✔ Text corpus: 5242249 characters, 1185427 tokens (roughly words)
#> 
#> ── Training model information ──────────────────────────────────────────────────
#> - Method:      FastText (Skip-Gram with Negative Sampling)
#> - Dimensions:  50
#> - Window size: 5 (5 words behind and 5 words ahead the current word)
#> - Subsampling: 1e-04
#> - Min. freq.:  5 occurrences in text
#> - Iterations:  5 training iterations
#> - CPU threads: 8
#> 
#> ── Training... 
#> ✔ Word vectors trained: 14207 unique tokens (time cost = 30 secs)

dt3
#> # wordvec (data.table): [14207 × 3] (normalized)
#>                word                     vec  freq
#>     1:          the [ 0.0168, ...<50 dims>] 58797
#>     2:          and [ 0.1486, ...<50 dims>] 32193
#>     3:            a [ 0.0195, ...<50 dims>] 31783
#>     4:           of [ 0.1441, ...<50 dims>] 29142
#>     5:           to [-0.0386, ...<50 dims>] 27218
#> ------                                           
#> 14203:        spray [ 0.1024, ...<50 dims>]     5
#> 14204: disabilities [ 0.1920, ...<50 dims>]     5
#> 14205:        crook [ 0.0948, ...<50 dims>]     5
#> 14206:     Syndrome [ 0.0086, ...<50 dims>]     5
#> 14207:      snipers [ 0.0958, ...<50 dims>]     5
most_similar(dt3, "Ive")  # evaluate performance
#> [Word Vector] =~ Ive
#> (normalized to unit length)
#>           word   cos_sim row_id
#>  1:       seen 0.8379643    110
#>  2:      Youve 0.8275185    945
#>  3:       Weve 0.8182750   3105
#>  4:      youve 0.7810110   3494
#>  5:      WORST 0.7722243   5898
#>  6:         ve 0.7701212   6816
#>  7:        ive 0.7497866   6913
#>  8:    funnier 0.7438126   7108
#>  9: beforehand 0.7387722   9171
#> 10:   Vonnegut 0.7381952  12894
most_similar(dt3, ~ man - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ man - he + she
#> (normalized to unit length)
#>        word   cos_sim row_id
#> 1:    woman 0.8673148    261
#> 2: salesman 0.7465448    299
#> 3:     girl 0.7362374   5594
#> 4:   madman 0.7354041   6553
#> 5: henchman 0.7347212  12263
most_similar(dt3, ~ boy - he + she, topn=5)  # evaluate performance
#> [Word Vector] =~ boy - he + she
#> (normalized to unit length)
#>       word   cos_sim row_id
#> 1:    girl 0.7968069    261
#> 2:   woman 0.7399783    299
#> 3:    boys 0.6797259    676
#> 4:     kid 0.6796938   1045
#> 5: younger 0.6494351   1248